% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{borgdorff_lLik}
\alias{borgdorff_lLik}
\title{#'Likelihood of Borgdorff Estimates
#'across the tb_progression groups
#'average at the end
#'@param Par_list list of parameters Mpfast, Mrslow, rfast, rRecov
#'@param n_red crude reductions the likelihood (allows non-sampling bias)
#'@return likelihood}
\usage{
borgdorff_lLik(Par_list, N_red = 1)
}
\description{
#'Likelihood of Borgdorff Estimates
#'across the tb_progression groups
#'average at the end
#'@param Par_list list of parameters Mpfast, Mrslow, rfast, rRecov
#'@param n_red crude reductions the likelihood (allows non-sampling bias)
#'@return likelihood
}
