% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{bspline}
\alias{bspline}
\title{This function is used to return the basis of a spline function.}
\usage{
bspline(x, k, i, m)
}
\value{
basis of a spline
}
\description{
This function is used to return the basis of a spline function.
}
