% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib.R
\name{calib}
\alias{calib}
\title{if any output is missing or negative or if any model state population is negative}
\usage{
calib(samp_i, optim_mat, loc, pdf = TRUE, cex.size = 1)
}
\description{
if any output is missing or negative or if any model state population is negative
}
