% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_reblnc_pop.R
\docType{data}
\name{can_go}
\alias{can_go}
\title{Rebalance to a New Distribution
First, setup a matrix of all allowable transitions between states.
For this model, individuals can only move a single step in either
direction, that is an individual in mortality level 2 can only move
to levels 1 or 3, but not to level 4, in a single time step.
The matrix can_go below captures these transition limitations. We also
create a matrix to hold the sum of transitions, did_go.}
\format{An object of class \code{matrix} with 16 rows and 16 columns.}
\usage{
can_go
}
\description{
Rebalance to a New Distribution
First, setup a matrix of all allowable transitions between states.
For this model, individuals can only move a single step in either
direction, that is an individual in mortality level 2 can only move
to levels 1 or 3, but not to level 4, in a single time step.
The matrix can_go below captures these transition limitations. We also
create a matrix to hold the sum of transitions, did_go.
}
\keyword{datasets}
