% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention_tttsetup.R
\name{create_ttt_dist}
\alias{create_ttt_dist}
\title{this function returns default values for ttt which would be total population}
\usage{
create_ttt_dist(ttt_list, results, PV)
}
\arguments{
\item{ttt_list}{vector of values from interface that define the intervention}

\item{results}{matrix of results (can be 1 entry from basecase)}

\item{PV}{formatted parameter vector}
}
\value{
ttt_params list of params for ttt interventions
}
\description{
this function returns default values for ttt which would be total population
}
