% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cSim}
\alias{cSim}
\title{cSim}
\usage{
cSim(nYrs, nRes, rDxt, TxQualt, InitPop, Mpfast, ExogInf, MpfastPI, Mrslow,
  rrSlowFB, rfast, RRcurDef, rSlfCur, p_HR, vTMort, RRmuRF, RRmuHR, muTbRF,
  Birthst, HrEntEx, ImmNon, ImmLat, ImmAct, ImmFst, mubt, RelInf, RelInfRg,
  Vmix, rEmmigFB, TxVec, TunTxMort, rDeft, rLtScrt, LtTxPar, LtDxPar, RRdxAge,
  rRecov, pImmScen, EarlyTrend, pReTx, NixTrans, dist_gen, trans_mat_tot_ages)
}
\arguments{
\item{nYrs}{number of years to run the model.}

\item{nRes}{number of results of the model}

\item{rDxt}{Rate of active TB diagnosis over time}

\item{TxQualt}{active TB treatment over time}

\item{InitPop}{Initial Population matrix}

\item{Mpfast}{Matrix of the probabilities of fast TB progression (age x TB prog risk group)}

\item{ExogInf}{Exogenous infection risk for non-US born population}

\item{MpfastPI}{Matrix of the probabilities of fast TB progression w/ partial immunity (age x TB prog risk group)}

\item{Mrslow}{matrix of the rates of slow progression (age x TB prog risk group)}

\item{rrSlowFB}{rate of fast TB progression}

\item{RRcurDef}{Rate Ratio for cure given treatment defaul}

\item{rSlfCur}{rate of self cure from active TB}

\item{p_HR}{probability of high risk population @ entry into the model}

\item{vTMort}{vector of TB mortality rates}

\item{RRmuRF}{rate ratio of mortality across mortality risk group}

\item{RRmuHR}{rate ratio of mortality across low/high risk dimension}

\item{Birthst}{Births over time}

\item{HrEntEx}{Matrix of Entry and Exit rates into the High Risk population}

\item{ImmNon}{Immigration with no TB}

\item{ImmLat}{Immigration with Latent TB}

\item{ImmAct}{Immigration with Active TB}

\item{ImmFst}{Immigration with Fast Progressing TB}

\item{mubt}{background mortality over time}

\item{RelInf}{beta}

\item{RelInfRg}{beta based off of risk group}

\item{Vmix}{1-sigma}

\item{rEmmigFB}{rate of emmigration in non-US born population}

\item{TxVec}{vector of parameters for TB Tx}

\item{TunTxMort}{Tuning parameter for mortality on TB Tx}

\item{rDeft}{rate of default from TB treatment over time}

\item{rLtScrt}{rate of latent screening over time}

\item{LtTxPar}{matrix of latent treatment parameters}

\item{LtDxPar}{matrix of latent diagnosis parameters}

\item{RRdxAge}{vector of rate ratios for TB diagnosis by age}

\item{rRecov}{rate of recovery from latent slow to safe tb state}

\item{pImmScen}{lack of reactivitiy to IGRA for Sp}

\item{EarlyTrend}{ramp down of TB in burn-in}

\item{pReTx}{probability of re-treatment for TB}

\item{dist_gen}{general distribution across tb progression and mort}

\item{muTBRF}{comorbidity factor btw TB and mortality risk group}
}
\value{
a list of outputs
}
\description{
runs a simulation of the tb model
}
