% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_reblnc_pop.R
\docType{data}
\name{cuts}
\alias{cuts}
\title{Define the number of cut points in the model; standard is 4}
\format{An object of class \code{numeric} of length 5.}
\usage{
cuts
}
\description{
Define the number of cut points in the model; standard is 4
}
\keyword{datasets}
