% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{dDirMult}
\alias{dDirMult}
\title{Calibration Functions for use with the tb_model.cpp model
This script creates several individual log likelihood functions
for the calibration of the State Level TB model in tb_model.cpp
These llikelihood functions are called in IMIS_functions.R
takes in the outputs and calibration data and creates likelihood functions
Dirichlet multinomial density function}
\usage{
dDirMult(M, n, Rho)
}
\arguments{
\item{M}{params of the dirichlet, i.e. model strain distribution}

\item{n}{category counts from the survey}

\item{rho}{correlation parameter = 1/sample size}
}
\value{
M density
}
\description{
Calibration Functions for use with the tb_model.cpp model
This script creates several individual log likelihood functions
for the calibration of the State Level TB model in tb_model.cpp
These llikelihood functions are called in IMIS_functions.R
takes in the outputs and calibration data and creates likelihood functions
Dirichlet multinomial density function
}
