% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention_costinputdefault.R
\name{def_costinputs}
\alias{def_costinputs}
\title{Function that will return default values of the program change values.}
\usage{
def_costinputs()
}
\value{
vector of the default values
}
\description{
Function that will return default values of the program change values.
}
