% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention_tttsetup.R
\name{def_ttt}
\alias{def_ttt}
\title{this function returns default values for ttt which would be total population}
\usage{
def_ttt()
}
\description{
this function returns default values for ttt which would be total population
}
