% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_PI.R
\name{dev_fin2_param_init}
\alias{dev_fin2_param_init}
\alias{dev_fin_param_init}
\title{This function takes the same inputs as the Outputs}
\usage{
dev_fin_param_init(PV, loc, prg_chng)
}
\arguments{
\item{PV}{vector of Inputs to format}

\item{loc}{two letter abbreviation}

\item{prg_chng}{vector of program change values}

\item{Int1}{boolean for intervention 1}

\item{Int2}{boolean for intervention 2}

\item{Int3}{boolean for intervention 3}

\item{Int4}{boolean for intervention 4}

\item{Int5}{boolean for intervention 5}

\item{Scen1}{boolean for scenario 1}

\item{Scen2}{boolean for scenario 2}

\item{Scen3}{boolean for scenario 3}

\item{ttt_list}{list of ttt changes}
}
\value{
InputParams list
}
\description{
This function takes the same inputs as the Outputs
}
