% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\docType{data}
\name{dist_gen}
\alias{dist_gen}
\title{Functions for likelihood of different published estimates}
\format{An object of class \code{matrix} with 4 rows and 4 columns.}
\usage{
dist_gen
}
\description{
Functions for likelihood of different published estimates
}
\keyword{datasets}
