% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_reblnc_pop.R
\docType{data}
\name{dist_goal}
\alias{dist_goal}
\title{Setup two new matrices for changes
The first matrix, dist_goal is the distribution we want, fed in as model input, a 4x1 vector
The second matrix, dist_orig is the starting distribution pulled from inside the timestep}
\format{An object of class \code{matrix} with 4 rows and 4 columns.}
\usage{
dist_goal
}
\description{
Setup two new matrices for changes
The first matrix, dist_goal is the distribution we want, fed in as model input, a 4x1 vector
The second matrix, dist_orig is the starting distribution pulled from inside the timestep
}
\keyword{datasets}
