% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_betterIMIS.R
\name{dpropnew}
\alias{dpropnew}
\title{This function ASSESS DENSITY OF PROPOSAL DISTRIBUTION}
\usage{
dpropnew(samps, mu, vcv, logg = F)
}
\value{
density
}
\description{
This function ASSESS DENSITY OF PROPOSAL DISTRIBUTION
}
