% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions_st.r
\name{dth_tot_lLik_st}
\alias{dth_tot_lLik_st}
\title{TOTAL US DEATHS
1970,1975,1980,1985,1990-2007
Motivation: norm, mean centered with CI = +/- 5% of mean}
\usage{
dth_tot_lLik_st(V, st)
}
\arguments{
\item{V}{table of deaths from 1970 to 2016}
}
\value{
likelihood
}
\description{
TOTAL US DEATHS
1970,1975,1980,1985,1990-2007
Motivation: norm, mean centered with CI = +/- 5% of mean
}
