% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{ferebee_lLik}
\alias{ferebee_lLik}
\title{Likelihood of Ferebee Estimates
across the tb_progression groups
average at the end}
\usage{
ferebee_lLik(Par, N_red = 4)
}
\arguments{
\item{Par_list}{list (pfast,rslow,rfast,rRecov)}
}
\value{
likelihood
}
\description{
Likelihood of Ferebee Estimates
across the tb_progression groups
average at the end
}
