% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_df.R
\name{format_df}
\alias{format_df}
\title{Formats Outputs to dataframe format for the TABBY2 interface}
\usage{
format_df(year, loc, nat, scen, int, df)
}
\arguments{
\item{loc}{location of results}

\item{nat}{nationality of results}

\item{scen}{scenario}

\item{int}{intervention}

\item{df}{dataframe of results}

\item{years}{vector of years}
}
\value{
csv of results formatted
}
\description{
Formats Outputs to dataframe format for the TABBY2 interface
}
