% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formatparaminit.R
\name{format_paramInit}
\alias{format_paramInit}
\title{takes a mean and interval and creates a distribution
fitting mean and the width of the interval
calculates the distribution values
might be unnecessary
calculating the transformed by the log or lgt
save it all}
\usage{
format_paramInit()
}
\description{
takes a mean and interval and creates a distribution
fitting mean and the width of the interval
calculates the distribution values
might be unnecessary
calculating the transformed by the log or lgt
save it all
}
