% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ParamInitst.R
\name{format_paramInitst}
\alias{format_paramInitst}
\title{takes a mean and interval and creates a distribution
fitting mean and the width of the interval
calculates the distribution values
might be unnecessary
calculating the transformed by the log or lgt
save it all}
\usage{
format_paramInitst()
}
\description{
takes a mean and interval and creates a distribution
fitting mean and the width of the interval
calculates the distribution values
might be unnecessary
calculating the transformed by the log or lgt
save it all
}
