% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_futurePC.R
\name{futurePC_graphs}
\alias{futurePC_graphs}
\title{graph of total diagnosed cases
by total population, US born population, and non-US born population
format the plot
plot the model data
multiply raw output by 1,000 to convert from millions to thousands
reported data for comparison
plot text
graph of total diagnosed cases
by total population, US born population, and non-US born population
format the plot
plot the model data
multiply raw output by 1,000 to convert from millions to thousands
reported data for comparison
plot text
Percent of Total Cases Non-US Born Population
format the plot
plot the model data
reported data for comparison
plot text
Percent of Non-US Born Cases from Recent Immigrant Population
format the plot
plot the model data
reported data for comparison
plot text
Treatment Outcomes 1993-2014
format the plot
plot the model data
reported data for comparison
plot text}
\usage{
futurePC_graphs(dfBc, dfP1, dfP2, dfP3, endyr)
}
\description{
graph of total diagnosed cases
by total population, US born population, and non-US born population
format the plot
plot the model data
multiply raw output by 1,000 to convert from millions to thousands
reported data for comparison
plot text
graph of total diagnosed cases
by total population, US born population, and non-US born population
format the plot
plot the model data
multiply raw output by 1,000 to convert from millions to thousands
reported data for comparison
plot text
Percent of Total Cases Non-US Born Population
format the plot
plot the model data
reported data for comparison
plot text
Percent of Non-US Born Cases from Recent Immigrant Population
format the plot
plot the model data
reported data for comparison
plot text
Treatment Outcomes 1993-2014
format the plot
plot the model data
reported data for comparison
plot text
}
