% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ParamInit.R
\name{gammapar}
\alias{gammapar}
\title{takes a mean and interval and creates a distribution
fitting mean and the width of the interval}
\usage{
gammapar(tgt)
}
\description{
takes a mean and interval and creates a distribution
fitting mean and the width of the interval
}
