% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_startval.R
\name{gen_st_val}
\alias{gen_st_val}
\title{Function to create a Starting Values for the Parameters
to be Optimized.}
\usage{
gen_st_val(n = 10, samp = "LHS")
}
\arguments{
\item{n}{number of datasets you want to generate}

\item{samp}{Which sampling method to use. "NORM" for random normal samples or "LHS" for latin hypercube sampling.}
}
\value{
rData file of starting values
}
\description{
Function to create a Starting Values for the Parameters
to be Optimized.
}
