% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_calib.R
\name{graphs_calib}
\alias{graphs_calib}
\alias{calib_graphs}
\title{This script creates graphs that compare model simulations
versus calibration data.}
\usage{
calib_graphs(df, Par_list)
}
\arguments{
\item{df}{dataframe of results}
}
\value{
.pdf of calibration graphs
}
\description{
This script creates graphs that compare model simulations
versus calibration data.
}
