% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_calib_st.R
\name{graphs_calib_st}
\alias{graphs_calib_st}
\alias{calib_graphs_st}
\title{This script creates graphs that compare model simulations
versus calibration data.}
\usage{
calib_graphs_st(df, loc, Par_list, pdf = TRUE, cex.size = 0.7)
}
\arguments{
\item{df}{dataframe of results}
}
\value{
.pdf of calibration graphs
}
\description{
This script creates graphs that compare model simulations
versus calibration data.
}
