% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{homeless_10_lLik}
\alias{homeless_10_lLik}
\title{Homeless Population in 2010
Motivation: normally distributed, mean centered with CI = +/- 25% of mean}
\usage{
homeless_10_lLik(V)
}
\arguments{
\item{V}{scalar value of homeless pop in 2010}
}
\value{
likelihood
}
\description{
Homeless Population in 2010
Motivation: normally distributed, mean centered with CI = +/- 25% of mean
}
