% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions_st.r
\docType{data}
\name{idx_5yr_end}
\alias{idx_5yr_end}
\title{Calibration Functions for use with the tb_model.cpp model for state level
This script creates several individual log likelihood functions
for the calibration of the State Level TB model in tb_model.cpp
These llikelihood functions are called in IMIS_functions.R
takes in the outputs and calibration data and creates likelihood functions}
\format{
An object of class \code{numeric} of length 5.
}
\usage{
idx_5yr_end
}
\description{
Calibration Functions for use with the tb_model.cpp model for state level
This script creates several individual log likelihood functions
for the calibration of the State Level TB model in tb_model.cpp
These llikelihood functions are called in IMIS_functions.R
takes in the outputs and calibration data and creates likelihood functions
}
\keyword{datasets}
