% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_betterIMIS.R
\name{Imis2}
\alias{Imis2}
\title{This function samples the prior using a random normal generation.}
\usage{
Imis2(
  B = 1000,
  B.re = 10000,
  number_k = 100,
  seed_val = 1,
  SaveNam = "xx",
  n_cores = 2,
  OptInit
)
}
\description{
This function samples the prior using a random normal generation.
}
