% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{invlgt}
\alias{invlgt}
\title{This function calculates the inverse logit of the value or parameter inputted.}
\usage{
invlgt(x)
}
\arguments{
\item{x}{value of interest, often an odds}
}
\value{
inverse logit of input x
}
\description{
This function calculates the inverse logit of the value or parameter inputted.
}
