% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{LgtCurve}
\alias{LgtCurve}
\title{This function is used to simulate a logistic increase between two years of interest
to a user-specified maximum. Useful for simulating slow increase/decrease of a
parameter over time.}
\usage{
LgtCurve(StYr, Endyr, EndVal)
}
\arguments{
\item{StYr}{Year to start the logit curve}

\item{Endyr}{Year to end the logit curve}

\item{EndVal}{maximum value that the logistic curve can reach}
}
\value{
vector of values demonstrating a logistic increase from zero to EndVal
}
\description{
This function is used to simulate a logistic increase between two years of interest
to a user-specified maximum. Useful for simulating slow increase/decrease of a
parameter over time.
}
