% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_IMIS_ag.R
\name{llikelihood_ag}
\alias{llikelihood_ag}
\title{Local parallelization via multicore}
\usage{
llikelihood_ag(start_mat, n_cores = 1)
}
\arguments{
\item{start_mat}{matrix of parameters}

\item{n_cores}{number of cores to use on the cluster}
}
\value{
lLik
}
\description{
Local parallelization via multicore
}
