% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_IMIS_demo.R
\name{llikelihood_demo}
\alias{llikelihood_demo}
\title{Local parallelization via multicore}
\usage{
llikelihood_demo(ParMatrix, n_cores = 1)
}
\arguments{
\item{ParMatrix}{matrix of parameters}

\item{n_cores}{number of cores to use on the cluster}
}
\value{
lLik
}
\description{
Local parallelization via multicore
}
