% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_IMIS_noTB.R
\name{llikelihood_noTB}
\alias{llikelihood_noTB}
\title{Local parallelization via multicore}
\usage{
llikelihood_noTB(start_mat, loc = "US", n_cores = 1)
}
\arguments{
\item{start_mat}{matrix of parameters}

\item{loc}{two digit postal code of location}

\item{n_cores}{number of cores to use on the cluster}
}
\value{
lLik
}
\description{
Local parallelization via multicore
}
