% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_IMIS_rblnc.R
\name{llikelihood_rblnc}
\alias{llikelihood_rblnc}
\title{Local parallelization via multicore}
\usage{
llikelihood_rblnc(start_mat, loc = "US", n_cores = 1)
}
\arguments{
\item{start_mat}{matrix of parameters}

\item{n_cores}{number of cores to use on the cluster}
}
\value{
lLik
}
\description{
Local parallelization via multicore
}
