% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_IMIS_demo.R
\name{llikelihoodZ_demo}
\alias{llikelihoodZ_demo}
\title{This scrprmst creates a function that loops over the log-likelihood
functions found in calib_functions.R and updates the
this is the function that goes into the optimizer}
\usage{
llikelihoodZ_demo(samp_i, ParMatrix)
}
\arguments{
\item{samp_i}{sample id}

\item{ParMatrix}{matrix of parameters  # Par = par_1}
}
\value{
lLik
load the necessary calibration data
Log-likelihood functions
Assign the calibration importance weights from CalibDat
These weights are based on year of the simulation.
format P
Log-likelihood functions
Assign the calibration importance weights from CalibDat
These weights are based on year of the simulation.
format P
if any output is missing or negative or if any model state population is negative
set the likelihood to a hugely negative number (penalized)
TOTAL POP EACH DECADE, BY US/FB
TOTAL POP AGE DISTRIBUTION 2016 index updated
Total DEATHS 1999-2016 BY AGE
Total DEATHS 1999-2016 BY AGE
}
\description{
This scrprmst creates a function that loops over the log-likelihood
functions found in calib_functions.R and updates the
this is the function that goes into the optimizer
}
