% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_IMIS_noRB.R
\name{llikelihoodZ_noRB}
\alias{llikelihoodZ_noRB}
\title{This scrprmst creates a function that loops over the log-likelihood
functions found in calib_functions.R and updates the
this is the function that goes into the optimizer}
\usage{
llikelihoodZ_noRB(samp_i, opt_mat)
}
\arguments{
\item{samp_i}{sample id}

\item{opt_mat}{matrix of parameters  # Par = par_1}
}
\value{
lLik
if any output is missing or negative or if any model state population is negative
set the likelihood to a hugely negative number (penalized)
#' TOTAL DIAGNOSED CASES 1953-2016 - index is same
v1   <- M[4:67,"NOTIF_ALL"]+M[4:67,"NOTIF_MORT_ALL"]
addlik <- notif_tot_lik(V=v1); addlik
lLik <- lLik + addlik
#' US CASES AGE DISTRIBUTION 1993-2016 - index updated
v2   <- M[44:67,205:215]+M[44:67,216:226]
v2a <- v2[,-11]; v2a[,10] <- v2a[,10]+v2[,11]
addlik <- notif_age_us_lLik(V=v2a); addlik
lLik <- lLik + addlik
#' FB CASES AGE DISTRIBUTION 1993-2016 - index updated
v2   <- (M[44:67,136:146]+M[44:67,189:199]) - (M[44:67,205:215]+M[44:67,216:226])
v2b <- v2[,-11]; v2b[,10] <- v2b[,10]+v2[,11]
addlik <- notif_age_fb_lLik(V=v2b); addlik
lLik <- lLik + addlik
#' CASES FB DISTRIBUTION 1993-2016 - index updated
v3   <- cbind(M[44:67,148]+M[44:67,149]+(M[44:67,201]+M[44:67,202]),
              M[44:67,147]+M[44:67,200])
addlik <- notif_fb_lLik(V=v3); addlik
lLik <- lLik + addlik
#' CASES FB, US 2010-2014  SLOPE - index updated
v3   <- cbind(M[44:66,148]+M[44:66,149]+(M[44:66,201]+M[44:66,202]),
              M[44:66,147]+M[44:66,200])
addlik <- notif_fbus_slp_lLik(V=v3); addlik
lLik <- lLik + addlik
#' CASES HR DISTRIBUTION 1993-2016 - index updated
v5b   <- cbind(M[44:67,151],M[44:67,150]) + cbind(M[44:67,204],M[44:67,203])
addlik <- notif_us_hr_lLik(V=v5b); addlik
lLik <- lLik + addlik
#' CASES FB RECENT ENTRY DISTRIBUTION 1993-2014 index updated
v6   <- M[44:65,148:149]+M[44:65,201:202]
addlik <- notif_fb_rec_lLik(V=v6); addlik
lLik <- lLik + addlik
#' TREATMENT OUTCOMES 1993-2012 - index updated
v11  <- M[44:65,132:134]
addlik <- tx_outcomes_lLik(V=v11); addlik
lLik <- lLik + addlik
#' TOTAL LTBI TREATMENT INITS 2002 - index updated
v12  <- M[53,152]
addlik <- tltbi_tot_lLik(V=v12); addlik
lLik <- lLik + addlik
#' DIST LTBI TREATMENT INITS 2002 - index updated
v13  <- M[53,153:154]/M[53,152]
addlik <- (tltbi_dist_lLik(V=v13))*2; addlik
lLik <- lLik + addlik
#' LTBI PREVALENCE BY AGE 2011, US - index updated
v15  <- cbind(M[62,55:65],M[62,33:43]-M[62,55:65])
v15a <- outer(v15[,1],c(P[["SensLt"]],1-P[["SensLt"]]))+outer(v15[,2],c(1-P[["SpecLt"]],P[["SpecLt"]]))
addlik <- ltbi_us_11_lLik(V=v15a)*2; addlik
lLik <- lLik + addlik
#' LTBI PREVALENCE BY AGE 2011, FB - index updated
v16  <- cbind(M[62,66:76],M[62,44:54]-M[62,66:76])
v16a <- outer(v16[,1],c(P[["SensLt"]],1-P[["SensLt"]]))+outer(v16[,2],c(1-P[["SpecLt"]],P[["SpecLt"]]))
addlik <- ltbi_fb_11_lLik(V=v16a)*2; addlik
#' lLik <- lLik + addlik
v17a  <- M[,31]+M[,32]
addlik <- tot_pop_yr_fb_lLik(V=v17a); addlik
lLik <- lLik + addlik
# v17b  <- M[,30]
# addlik <- tot_pop_yr_usb_lLik(V=v17b); addlik
# lLik <- lLik + addlik
#' TOTAL POP AGE DISTRIBUTION 2016 index updated
v18  <- cbind(M[67,33:43],M[67,44:54])
addlik <- tot_pop16_ag_fb_lLik(V=v18); addlik
lLik <- lLik + addlik
  Total DEATHS by Decade
#' Total DEATHS 1999-2016 BY AGE

#' HOMELESS POP 2010 - index updated
v23b  <- M[61,29]
addlik <- homeless_10_lLik(V=v23b); addlik
lLik <- lLik + addlikd
LIKELIHOOD FOR BORGDORFF, FEREBEE & SUTHERLAND ESTIMATES
}
\description{
This scrprmst creates a function that loops over the log-likelihood
functions found in calib_functions.R and updates the
this is the function that goes into the optimizer
}
