% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_IMIS_noTB_rblnc.R
\name{llikelihoodZ_noTB_rblnc}
\alias{llikelihoodZ_noTB_rblnc}
\title{This scrprmst creates a function that loops over the log-likelihood
functions found in calib_functions.R and updates the
this is the function that goes into the optimizer}
\usage{
llikelihoodZ_noTB_rblnc(samp_i, opt_mat)
}
\arguments{
\item{samp_i}{sample id}

\item{opt_mat}{matrix of parameters  # Par = par_1}
}
\value{
lLik
format P
if any output is missing or negative or if any model state population is negative
set the likelihood to a hugely negative number (penalized)
TOTAL POP AGE DISTRIBUTION 2016 index updated
  Total DEATHS by Decade
#' Total DEATHS 1999-2016 BY AGE

#' Mort_dist 2016
}
\description{
This scrprmst creates a function that loops over the log-likelihood
functions found in calib_functions.R and updates the
this is the function that goes into the optimizer
}
