% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_IMIS_noTB.R
\name{llikelihoodZ_noTB_st}
\alias{llikelihoodZ_noTB_st}
\title{This scrprmst creates a function that loops over the log-likelihood
functions found in calib_functions.R and updates the
this is the function that goes into the optimizer}
\usage{
llikelihoodZ_noTB_st(samp_i, opt_mat, loc)
}
\arguments{
\item{samp_i}{sample id}

\item{opt_mat}{matrix of parameters  # Par = par_1}
}
\value{
lLik
format P
if any output is missing or negative or if any model state population is negative
set the likelihood to a hugely negative number (penalized)
Total DEATHS 1979-2016
Total DEATHS by Decade
#' Total DEATHS 2015-2016 BY AGE
#' Mort_dist 2016
}
\description{
This scrprmst creates a function that loops over the log-likelihood
functions found in calib_functions.R and updates the
this is the function that goes into the optimizer
}
