% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_optim.R
\name{load_optim}
\alias{load_optim}
\title{loads data for the geography of interest}
\usage{
load_optim()
}
\value{
loc_vec of optimized geographies
}
\description{
loads data for the geography of interest
}
