% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tabby.R
\name{load_tabby}
\alias{load_tabby}
\title{loads data for the geography of interest}
\usage{
load_tabby(loc = "US")
}
\arguments{
\item{loc}{two letter postal abbreviation for states; US for national}
}
\value{
env
create a new environment
load necessary datasets into the environment
}
\description{
loads data for the geography of interest
}
