% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{logit}
\alias{logit}
\alias{lgt}
\title{lgt}
\usage{
lgt(x)
}
\arguments{
\item{x}{value of interest, often an odds}
}
\value{
logit of input x
}
\description{
This function calculates the logit of the value or parameter inputted.
}
