% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prior.R
\name{lprior}
\alias{lprior}
\title{calculates the prior value for a vector of parameters}
\usage{
lprior(ParMatrix = ParInit)
}
\arguments{
\item{ParMatrix}{matrix of Initial parameters}
}
\value{
lPri
}
\description{
calculates the prior value for a vector of parameters
}
