% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_priorfunction.R
\name{lPrior2}
\alias{lPrior2}
\title{calculates the parameter prior with Jacobians}
\usage{
lPrior2(Par, Par3)
}
\arguments{
\item{Par}{parameters translated to the normal space}

\item{Par3}{parameters in their original distributions}
}
\value{
ldensity3
}
\description{
calculates the parameter prior with Jacobians
}
