% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{ltbi_fb_11_dp_lLik}
\alias{ltbi_fb_11_dp_lLik}
\title{Likelihood function for double positive LTBI in the NUS born based on 2011 NHANES}
\usage{
ltbi_fb_11_dp_lLik(V)
}
\arguments{
\item{V}{LTBI in FB pop 2011 (row=11 ages, col= ltbi, non-ltbi)}
}
\value{
likelihood
}
\description{
Likelihood function for double positive LTBI in the NUS born based on 2011 NHANES
}
