% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{ltbi_fb_11_lLik}
\alias{ltbi_fb_11_lLik}
\title{LTBI PREVALENCE BY AGE 2011, FB
Motivation: multinomial adjusted to match effective sample size due to survey weighting}
\usage{
ltbi_fb_11_lLik(V)
}
\arguments{
\item{V}{LTBI in FB pop 2011 (row=11 ages, col= ltbi, non-ltbi)}
}
\value{
likelihood
}
\description{
LTBI PREVALENCE BY AGE 2011, FB
Motivation: multinomial adjusted to match effective sample size due to survey weighting
}
