% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{ltbi_us_11_lLik}
\alias{ltbi_us_11_lLik}
\title{LTBI PREVALENCE BY AGE 2011, US
Motivation: additional prior on LTBI, using beta densities parameterized to Miramontes/Hill results}
\usage{
ltbi_us_11_lLik(V)
}
\arguments{
\item{V}{LTBI in US pop 2011 (row=11 ages, col= ltbi, non-ltbi)}
}
\value{
likelihood
}
\description{
LTBI PREVALENCE BY AGE 2011, US
Motivation: additional prior on LTBI, using beta densities parameterized to Miramontes/Hill results
}
