% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_allscenarios.R, R/model_calib_outputs.R
\name{make_all_scenarios}
\alias{make_all_scenarios}
\alias{model_calib_outputs}
\title{THIS FUNCTION INPUTS A TABLE OF PARAMETERS AND RUNS THE TB MODEL
AND GENERATES AN ARRAY OF OUTPUTS FOR EACH POSSIBLE PREDEFINED
SCENARIO; THIS ARRAY IS THEN SAVED TO THE DATA FOLDER FOR THAT
LOCATION}
\usage{
make_all_scenarios(loc, ParMatrix)

model_calib_outputs(loc = "US", bc.array, samp_i = 1, simp.date)
}
\arguments{
\item{loc}{two character location code for the model location}

\item{ParMatrix}{parameters to use in the simulation}

\item{bc.array}{base case results array}

\item{simp.date}{date to append to the files; should match the optim date}

\item{loc}{two character location code for the model location}

\item{samp.i}{which simulation to use (1:10)}
}
\description{
THIS FUNCTION INPUTS A TABLE OF PARAMETERS AND RUNS THE TB MODEL
AND GENERATES AN ARRAY OF OUTPUTS FOR EACH POSSIBLE PREDEFINED
SCENARIO; THIS ARRAY IS THEN SAVED TO THE DATA FOLDER FOR THAT
LOCATION

THIS FUNCTION READS IN THE BASE CASE SCENARIO AND THEN
CREATES INDIVIDUAL RDS FILES FOR OUTPUTS NECESSARY FOR
THE TABBY2 COMPARISON TO RECENT DATA PLOTS
}
