% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabby2_caliboutputs.r
\name{model_calib_outputs}
\alias{model_calib_outputs}
\title{THIS FUNCTION READS IN THE BASE CASE SCENARIO AND THEN
CREATES INDIVIDUAL RDS FILES FOR OUTPUTS NECESSARY FOR
THE TABBY2 COMPARISON TO RECENT DATA PLOTS}
\usage{
model_calib_outputs(
  loc = "US",
  bc.array,
  samp_i = 1,
  simp.date,
  out_dir = file.path("~/MITUS/inst", loc, "calibration_outputs")
)
}
\arguments{
\item{loc}{two character location code for the model location}

\item{bc.array}{base case results array}

\item{simp.date}{date to append to the files; should match the optim date}

\item{out_dir}{directory the files are written to; defaults to the location's
calibration_outputs folder in a ~/MITUS checkout}

\item{samp.i}{which simulation to use (1:10)}
}
\description{
THIS FUNCTION READS IN THE BASE CASE SCENARIO AND THEN
CREATES INDIVIDUAL RDS FILES FOR OUTPUTS NECESSARY FOR
THE TABBY2 COMPARISON TO RECENT DATA PLOTS
}
