% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_load.R
\name{model_load_demo}
\alias{model_load_demo}
\title{loads data for the geography of interest}
\usage{
model_load_demo(loc = "US")
}
\arguments{
\item{loc}{two letter postal abbreviation for states; US for national}
}
\value{
void
lazy load necessary datasets
Model Input
creation of background parameters
elements of P will be replaced from either the StartVals in the case
of optimization or the user inputted dataset

format the calibration data
}
\description{
loads data for the geography of interest
}
