% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{mort_dist_lLik}
\alias{mort_dist_lLik}
\title{Mortality Risk Group Distribution 1999-2014
Motivation: dirichlet-multinomial, multinomial data with additional non-sampling biases}
\usage{
mort_dist_lLik(V, ESS = 500)
}
\arguments{
\item{V}{2016 distribution}

\item{ESS}{correlation parameter}
}
\value{
likelihood
}
\description{
Mortality Risk Group Distribution 1999-2014
Motivation: dirichlet-multinomial, multinomial data with additional non-sampling biases
}
