% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions_st.r
\name{mort_dist_lLik_norm_st}
\alias{mort_dist_lLik_norm_st}
\title{Mortality Risk Group Distribution 1999-2014
Motivation: dnorm}
\usage{
mort_dist_lLik_norm_st(V)
}
\arguments{
\item{V}{table of mort_dist 1999-2014 (row=16 years, col=11 ages)}

\item{rho}{correlation parameter}
}
\value{
likelihood
}
\description{
Mortality Risk Group Distribution 1999-2014
Motivation: dnorm
}
