% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_simulations.R
\name{national_OutputsZint}
\alias{national_OutputsZint}
\title{THIS FUNCTION INPUTS A TABLE OF PARAMETERS AND RUNS THE TB MODEL
AND GENERATES AN ARRAY OF OUTPUTS}
\usage{
national_OutputsZint(
  samp_i = 1,
  ParMatrix,
  loc,
  output_month = 11,
  startyr = 1950,
  endyr = 2050,
  Int1 = 0,
  Int2 = 0,
  Int3 = 0,
  Int4 = 0,
  Int5 = 0,
  Scen1 = 0,
  Scen2 = 0,
  Scen3 = 0,
  Scen4 = 0,
  Scen5 = 0,
  Scen6 = 0,
  prg_chng = def_prgchng(Par[1, ]),
  ttt_list = list(def_ttt_nat_ag()),
  care_cascade = def_care_cascade(),
  par2020 = parCOVID,
  return_params = def_returnScenario()
)
}
\arguments{
\item{samp_i}{which row of the parameter matrix to use}

\item{ParMatrix}{parameter matrix to use in the simulation}

\item{loc}{two-digit abbreviation for location}

\item{output_month}{when to output results}

\item{startyr}{year to start the simulation}

\item{endyr}{year to end the simulation}

\item{Int1}{boolean for intervention 1}

\item{Int2}{boolean for intervention 2}

\item{Int3}{boolean for intervention 3}

\item{Int4}{boolean for intervention 4}

\item{Int5}{boolean for intervention 5}

\item{Scen1}{boolean for scenario 1}

\item{Scen2}{boolean for scenario 2}

\item{Scen3}{boolean for scenario 3}

\item{Scen4}{boolean for scenario 4}

\item{Scen5}{boolean for scenario 3}

\item{Scen6}{boolean for scenario 4}

\item{prg_chng}{vector of program change values}

\item{ttt_list}{list of targeted testing and treatment values}

\item{care_cascade}{list of parameters for changes to LTBI care cascade}

\item{par2020}{vector of 2020 adjustment parameters}

\item{return_params}{list of parameters to govern the transition to post-COVID param levels}
}
\value{
results data frame of output
}
\description{
THIS FUNCTION INPUTS A TABLE OF PARAMETERS AND RUNS THE TB MODEL
AND GENERATES AN ARRAY OF OUTPUTS
}
