% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_initial.R
\name{national_param_init}
\alias{national_param_init}
\title{This function takes the same inputs as the Outputs}
\usage{
national_param_init(
  PV,
  loc,
  Int1 = 0,
  Int2 = 0,
  Int3 = 0,
  Int4 = 0,
  Int5 = 0,
  Scen1 = 0,
  Scen2 = 0,
  Scen3 = 0,
  prg_chng,
  ttt_list,
  delay = 0,
  immig2020Vec = c(1, 1, 1, 1, 1, 1),
  return_months = 865:888,
  multiplier = 1
)
}
\arguments{
\item{PV}{vector of Inputs to format}

\item{loc}{two letter abbreviation}

\item{Int1}{boolean for intervention 1}

\item{Int2}{boolean for intervention 2}

\item{Int3}{boolean for intervention 3}

\item{Int4}{boolean for intervention 4}

\item{Int5}{boolean for intervention 5}

\item{Scen1}{boolean for scenario 1}

\item{Scen2}{boolean for scenario 2}

\item{Scen3}{boolean for scenario 3}

\item{prg_chng}{vector of program change values}

\item{ttt_list}{list of ttt changes}

\item{immig2020Vec}{vector of immigration adjustment parameters}

\item{return_months}{vector of months to transition to post COVID-19 level}

\item{multiplier}{scalar multiplier for return immigration}
}
\value{
InputParams list
}
\description{
This function takes the same inputs as the Outputs
}
