% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_param.R
\name{new_param}
\alias{new_param}
\title{THE CODE BELOW SOURCES THE MODEL INPUTS AND THEN FORMATS THEM FOR USE
IN THE OUTPUTSINTZ FUNCTION THAT CALLS CSIM FROM THE TB_MODEL.CPP
FUNCTION FILE. ALL VARIABLE NAMES THAT END IN t ARE INDEXED BY TIME
VARIABLE NAMES BEGINNING WITH m ARE MATRICES & V ARE VECTORS.}
\usage{
new_param(PV, prg_chng)
}
\arguments{
\item{prg_chng}{vector of program change values}

\item{P}{vector of}
}
\value{
Params list
}
\description{
THE CODE BELOW SOURCES THE MODEL INPUTS AND THEN FORMATS THEM FOR USE
IN THE OUTPUTSINTZ FUNCTION THAT CALLS CSIM FROM THE TB_MODEL.CPP
FUNCTION FILE. ALL VARIABLE NAMES THAT END IN t ARE INDEXED BY TIME
VARIABLE NAMES BEGINNING WITH m ARE MATRICES & V ARE VECTORS.
}
