% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{notif_age_fb_lLik}
\alias{notif_age_fb_lLik}
\title{FB CASES AGE DISTRIBUTION 1993-2013
Motivation: dirichlet-multinomial, multinomial data with additional non-sampling biases}
\usage{
notif_age_fb_lLik(V, rho = 0.001)
}
\arguments{
\item{V}{table of fb notifications by age 1993-2013 (row=21 years, col=11 ages)}

\item{rho}{correlation parameter}
}
\value{
likelihood
}
\description{
FB CASES AGE DISTRIBUTION 1993-2013
Motivation: dirichlet-multinomial, multinomial data with additional non-sampling biases
}
