% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{notif_age_us_lLik}
\alias{notif_age_us_lLik}
\title{US Cases Age Distribution 1993-2016
Motivation: dirichlet-multinomial data with additional non-sampling biases}
\usage{
notif_age_us_lLik(V, rho = 0.005)
}
\arguments{
\item{V}{table of us notifications by age 1993-2013 (row=21 years, col=11 ages)}

\item{rho}{correlation parameter}
}
\value{
likelihood
}
\description{
US Cases Age Distribution 1993-2016
Motivation: dirichlet-multinomial data with additional non-sampling biases
}
