% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions_st.r
\name{notif_fb_5yr_lik}
\alias{notif_fb_5yr_lik}
\title{FB Diagnosed Cases 1953-2016
Motivation: Normal, mean centered with CI = +/- 5% of the mean}
\usage{
notif_fb_5yr_lik(V, st = st)
}
\arguments{
\item{V}{vector of total notifications 1953-2014}
}
\value{
likelihood
}
\description{
FB Diagnosed Cases 1953-2016
Motivation: Normal, mean centered with CI = +/- 5% of the mean
}
