% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{notif_fb_lLik}
\alias{notif_fb_lLik}
\title{CASES FB DISTRIBUTION 1993-2014
Motivation: dirichlet-multinomial, multinomial data with additional non-sampling biases#'@param V table of notifications by fb 1993-2014 (row=22 years, col=fb then us)}
\usage{
notif_fb_lLik(V, rho = 0.005)
}
\arguments{
\item{rho}{correlation parameter}
}
\value{
likelihood
}
\description{
CASES FB DISTRIBUTION 1993-2014
Motivation: dirichlet-multinomial, multinomial data with additional non-sampling biases#'@param V table of notifications by fb 1993-2014 (row=22 years, col=fb then us)
}
