% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{notif_fb_rec_lLik}
\alias{notif_fb_rec_lLik}
\title{CASES FB RECENT ENTRY DISTRIBUTION 1993-2013
Motivation: dirichlet-multinomial, multinomial data with additional non-sampling biases}
\usage{
notif_fb_rec_lLik(V, rho = 0.005)
}
\arguments{
\item{V}{table of notifications by FB 1993-2014 (row=22 years, col=pos then neg)}

\item{rho}{correlation parameter}
}
\value{
likelihood
}
\description{
CASES FB RECENT ENTRY DISTRIBUTION 1993-2013
Motivation: dirichlet-multinomial, multinomial data with additional non-sampling biases
}
