% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{notif_hr2_lLik}
\alias{notif_hr2_lLik}
\title{CASES HR DISTRIBUTION 1993-2014
Motivation: dirichlet-multinomial, multinomial data with additional non-sampling biases}
\usage{
notif_hr2_lLik(V, rho = 0.1)
}
\arguments{
\item{V}{table of notifications by hr 1993-2014 (row=22 years, col=pos then neg)}

\item{rho}{correlation parameter}
}
\value{
likelihood
}
\description{
CASES HR DISTRIBUTION 1993-2014
Motivation: dirichlet-multinomial, multinomial data with additional non-sampling biases
}
