% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{notif_hr_dist_lLik}
\alias{notif_hr_dist_lLik}
\title{CASES HR NATIVITY DISTRIBUTION 1993-2018}
\usage{
notif_hr_dist_lLik(V, rho = 0.05)
}
\value{
likelihood
}
\description{
CASES HR NATIVITY DISTRIBUTION 1993-2018
}
