% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txhistory.R
\name{notif_prev_lLik}
\alias{notif_prev_lLik}
\title{CASES TX HISTORY DISTRIBUTION 1993-2013
Motivation: dirichlet-multinomial, multinomial data with additional non-sampling biases}
\usage{
notif_prev_lLik(V, rho = 0.005)
}
\arguments{
\item{V}{table of notifications by tx history 1993-2013 (row=21 years, col=n then e)}
}
\value{
likelihood
}
\description{
CASES TX HISTORY DISTRIBUTION 1993-2013
Motivation: dirichlet-multinomial, multinomial data with additional non-sampling biases
}
