% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\docType{data}
\name{notif_tot}
\alias{notif_tot}
\title{Total Diagnosed Cases 1953-2016
Motivation: Normal, mean centered with CI = +/- 5% of the mean}
\format{An object of class \code{numeric} of length 64.}
\usage{
notif_tot
}
\arguments{
\item{V}{vector of total notifications 1953-2014}
}
\value{
likelihood
}
\description{
Total Diagnosed Cases 1953-2016
Motivation: Normal, mean centered with CI = +/- 5% of the mean
}
\keyword{datasets}
