% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions_st.r
\name{notif_tot_lLik_st}
\alias{notif_tot_lLik_st}
\title{Total Diagnosed Cases 1953-2016
Motivation: Normal, mean centered with CI = +/- 5% of the mean}
\usage{
notif_tot_lLik_st(V, st)
}
\arguments{
\item{V}{vector of total notifi
cations 1953-2014}
}
\value{
likelihood
}
\description{
Total Diagnosed Cases 1953-2016
Motivation: Normal, mean centered with CI = +/- 5% of the mean
}
