% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{notif_us_lik}
\alias{notif_us_lik}
\title{US Diagnosed Cases 1953-2016
Motivation: Normal, mean centered with CI = +/- 5% of the mean}
\usage{
notif_us_lik(V)
}
\arguments{
\item{V}{vector of total notifications 1953-2014}
}
\value{
likelihood
}
\description{
US Diagnosed Cases 1953-2016
Motivation: Normal, mean centered with CI = +/- 5% of the mean
}
