% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_optim_ag.R
\name{optim_ag}
\alias{optim_ag}
\title{This function is used to run a optimization on the input parameters.
when there is no rebalancing in the model.
It will return 7 datasets of the optimized parameters --one from each
optimization step.
The function requires the use of the ParamInit Rdata file and the
StartValues Rdata file.}
\usage{
optim_ag(df, samp_i = 1)
}
\arguments{
\item{df}{dataframe or matrix of starting values data frame}

\item{samp_i}{which rows of the data frame to use}
}
\value{
7 datasets from optimization loop
}
\description{
This function is used to run a optimization on the input parameters.
when there is no rebalancing in the model.
It will return 7 datasets of the optimized parameters --one from each
optimization step.
The function requires the use of the ParamInit Rdata file and the
StartValues Rdata file.
}
