% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_load.R
\name{optim_load}
\alias{optim_load}
\title{loads data for the geography of interest}
\usage{
optim_load()
}
\value{
loc_vec of optimized geographies
}
\description{
loads data for the geography of interest
}
