% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_optim_to_tabby2.R
\name{optim_to_tabby2}
\alias{optim_to_tabby2}
\title{THIS FUNCTION READS IN THE OPTIM FILE THAT MATCHES THE SIMP.DATE
INPUT AND CREATES ALL NECESSARY FILES FOR TABBY2 FOR THAT LOCATION.
ALL PREDEFINED SCENARIOS (RESHAPED AND ORIGINAL FILES) AND MODEL
CALIBRATION OUTPUTS. IT ALSO REMOVES ALL OLD CALIBRATION OUTPUTS.}
\usage{
optim_to_tabby2(loc, simp.date = "724")
}
\arguments{
\item{loc}{two character location code for the model location}

\item{simp.date}{date to append to the files; should match the optim date}
}
\description{
THIS FUNCTION READS IN THE OPTIM FILE THAT MATCHES THE SIMP.DATE
INPUT AND CREATES ALL NECESSARY FILES FOR TABBY2 FOR THAT LOCATION.
ALL PREDEFINED SCENARIOS (RESHAPED AND ORIGINAL FILES) AND MODEL
CALIBRATION OUTPUTS. IT ALSO REMOVES ALL OLD CALIBRATION OUTPUTS.
}
