% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{ORAdd}
\alias{ORAdd}
\title{Basic Functions Used In Package
This script includes several functions that are useful for data manipulation
throughout the package.
The first function takes the odds of a risk, then calculates the join
probability of that odds and and an inputted Odds Ratio that the user inputs, then
generates the corresponding risk.}
\usage{
ORAdd(val, OR)
}
\arguments{
\item{val}{known risk for population of interest}

\item{OR}{odds ratio by which the known risk will be modified}
}
\value{
new risk for population
}
\description{
Basic Functions Used In Package
This script includes several functions that are useful for data manipulation
throughout the package.
The first function takes the odds of a risk, then calculates the join
probability of that odds and and an inputted Odds Ratio that the user inputs, then
generates the corresponding risk.
}
