% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_sims.R
\name{OutputsZint}
\alias{OutputsZint}
\title{THIS FUNCTION INPUTS A TABLE OF PARAMETERS AND RUNS THE TB MODEL
AND GENERATES AN ARRAY OF OUTPUTS
load("data/ParamInit_2018.rData")
function to run the model}
\usage{
OutputsZint(samp_i = 1, ParMatrix, endyr = 2050, Int1 = 0, Int2 = 0,
  Int3 = 0, Int4 = 0, Int5 = 0, Scen1 = 0, Scen2 = 0, Scen3 = 0)
}
\arguments{
\item{samp_i}{how many samples}

\item{ParMatrix}{parameters to use in the simulation}

\item{endyr}{year to end the simulation}

\item{Int1}{boolean for intervention 1}

\item{Int2}{boolean for intervention 2}

\item{Int3}{boolean for intervention 3}

\item{Int4}{boolean for intervention 4}

\item{Int5}{boolean for intervention 5}

\item{Scen1}{boolean for scenario 1}

\item{Scen2}{boolean for scenario 2}

\item{Scen3}{boolean for scenario 3}
}
\value{
results data frame of output
}
\description{
THIS FUNCTION INPUTS A TABLE OF PARAMETERS AND RUNS THE TB MODEL
AND GENERATES AN ARRAY OF OUTPUTS
load("data/ParamInit_2018.rData")
function to run the model
}
