% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_load.R
\docType{data}
\name{P}
\alias{P}
\title{creation of background parameters
elements of P will be replaced from either the StartVals in the case
of optimization or the user inputted dataset}
\format{An object of class \code{numeric} of length 69.}
\usage{
P
}
\description{
creation of background parameters
elements of P will be replaced from either the StartVals in the case
of optimization or the user inputted dataset
}
\keyword{datasets}
