% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ParamInit}
\alias{ParamInit}
\title{Let's try documenting something}
\format{An object of class \code{data.frame} with 69 rows and 8 columns.}
\usage{
ParamInit
}
\description{
Let's try documenting something
}
\keyword{datasets}
