% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_reblnc_pop.R
\docType{data}
\name{params}
\alias{params}
\title{Define the size of the target population;
Define the distribution of the targeted population}
\format{An object of class \code{numeric} of length 3.}
\usage{
params
}
\description{
Define the size of the target population;
Define the distribution of the targeted population
}
\keyword{datasets}
