% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_reblnc_pop.R
\docType{data}
\name{pars}
\alias{pars}
\title{Use the pmvnorm function to computer the distribution function of the multivariate normal
distribution.}
\format{An object of class \code{numeric} of length 3.}
\usage{
pars
}
\description{
Use the pmvnorm function to computer the distribution function of the multivariate normal
distribution.
}
\keyword{datasets}
