% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_reblnc_pop.R
\docType{data}
\name{pars1}
\alias{pars1}
\title{Change distribution
Parameters decided by function above;}
\format{An object of class \code{numeric} of length 3.}
\usage{
pars1
}
\description{
Change distribution
Parameters decided by function above;
}
\keyword{datasets}
