% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_func.R
\name{posterior}
\alias{posterior}
\title{This function calculates the posterior using the
lprior and llikelihood functions.}
\usage{
posterior(theta)
}
\arguments{
\item{theta}{vector of values}

\item{n_cores}{number of cores to use for the optimization}
}
\value{
posterior
}
\description{
This function calculates the posterior using the
lprior and llikelihood functions.
}
