% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prgm_chngs.R
\name{prg_chg}
\alias{prg_chg}
\alias{prm_chg}
\title{the programatic changes provided by tabbyII are:
LTBI treatment cascade:
1.Screening coverage for currently defined risk groups (Multiple of current coverage, from 1.0 to 5.0 times)
2.Fraction receiving IGRA (% from 0 to 100) (rest assumed to get TST)
3.Fraction of individuals testing positive who accept treatment (0% to 100%)
4.Fraction of individuals initiating treatment who complete treatment (0% to 100%)
5.Treatment efficacy for individuals completing treatment (0% to 100%)}
\usage{
prm_chg(P, UI_prgm_chng)
}
\arguments{
\item{P}{vector of baseline Parameter Values}

\item{UI_prgm_chng}{vector of UI values for program changes}
}
\value{
P vector of updated baseline Parameter Values
}
\description{
TB Treatment Cascade
1.Average time to treatment for incident TB cases (0-100% of current value)
2.Fraction discontinuing/defaulting from treatment (0% to 100%)
}
