% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reblncd}
\alias{reblncd}
\title{reblncd}
\usage{
reblncd(mubt, can_go, RRmuHR, RRmuRF, HRdist, dist_gen_v)
}
\arguments{
\item{IP}{dataframe of formatted parameters}
}
\value{
trans_mat_tot_ages
}
\description{
this function calculates a matrix of transition probabilities for rebalancing in the model.
}
