% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{recent_trans_dist_lLik}
\alias{recent_trans_dist_lLik}
\title{DISTRIBUTION OF CASES RECENT TRANSMISSION VS NO RECENT TRANSMISSION}
\usage{
recent_trans_dist_lLik(V)
}
\arguments{
\item{V}{distribution of recent and not recent transmission}
}
\value{
likelihood
}
\description{
DISTRIBUTION OF CASES RECENT TRANSMISSION VS NO RECENT TRANSMISSION
}
