% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions_st.r
\name{recent_trans_dist_lLik_st}
\alias{recent_trans_dist_lLik_st}
\title{DISTRIBUTION OF CASES RECENT TRANSMISSION VS NO RECENT TRANSMISSION}
\usage{
recent_trans_dist_lLik_st(V, st)
}
\arguments{
\item{V}{distribution of recent and not recent transmission}
}
\value{
likelihood
}
\description{
DISTRIBUTION OF CASES RECENT TRANSMISSION VS NO RECENT TRANSMISSION
}
