% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_results.R, R/reshape_results.R
\name{reshape_results}
\alias{reshape_results}
\title{THIS FUNCTION INPUTS A TABLE OF PARAMETERS AND RUNS THE TB MODEL
AND GENERATES AN ARRAY OF OUTPUTS FOR EACH POSSIBLE PREDEFINED
SCENARIO; THIS ARRAY IS THEN SAVED TO THE DATA FOLDER FOR THAT
LOCATION}
\usage{
reshape_results(loc, results)

reshape_results(loc, results)
}
\arguments{
\item{loc}{two character location code for the model location}

\item{results}{is a list of results to be reshaped}

\item{loc}{two character location code for the model location}
}
\description{
THIS FUNCTION INPUTS A TABLE OF PARAMETERS AND RUNS THE TB MODEL
AND GENERATES AN ARRAY OF OUTPUTS FOR EACH POSSIBLE PREDEFINED
SCENARIO; THIS ARRAY IS THEN SAVED TO THE DATA FOLDER FOR THAT
LOCATION

THIS FUNCTION RESHAPES THE 9 INDIVIDUAL FILES INTO THE
FORMAT NEEDED FOR TABBY2 VISUALIZATION
}
