% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_run.R
\name{run_model}
\alias{run_model}
\title{load the necessary libraries
if any output is missing or negative or if any model state population is negative
set the likelihood to a hugely negative number (penalized)}
\usage{
run_model(samp_i, ParMatrix)
}
\description{
load the necessary libraries
if any output is missing or negative or if any model state population is negative
set the likelihood to a hugely negative number (penalized)
}
