% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_sampleprior.R
\name{sample.prior}
\alias{sample.prior}
\title{This function samples the prior using Latin Hypercube sampling method.}
\usage{
sample.prior(n)
}
\arguments{
\item{n}{number of sample sets to generate}
}
\value{
n sample sets from the prior in a single object
}
\description{
This function samples the prior using Latin Hypercube sampling method.
}
