% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMIS_functions.R
\name{sample.prior1}
\alias{sample.prior1}
\title{sample the prior function}
\usage{
sample.prior1(n)
}
\arguments{
\item{n}{n of the sample}
}
\value{
random sample of the prior
}
\description{
sample the prior function
}
