% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_results.R
\name{save.results}
\alias{save.results}
\title{save.results}
\usage{
save.results(df)
}
\arguments{
\item{df}{dataframe of output}
}
\value{
.csv of results (unformatted)
}
\description{
function to save results
}
