% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_script.R
\name{ScriptMak}
\alias{ScriptMak}
\title{Function to create scripts for optimization}
\usage{
ScriptMak(b)
}
\arguments{
\item{b}{number of scripts to generate}
}
\value{
b number of scripts for optimization
}
\description{
Function to create scripts for optimization
}
