% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{SmoCurve}
\alias{SmoCurve}
\title{This function creates a smooth curve from a vector of values.}
\usage{
SmoCurve(vec)
}
\arguments{
\item{vec}{vector of values that the user would like to have smoothed into a curve}
}
\value{
expanded vector of values that determine the shape of the smoothing spline
}
\description{
This function creates a smooth curve from a vector of values.
}
