% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{SmoCurve_decade}
\alias{SmoCurve_decade}
\title{This function creates a smooth curve from a vector of values.}
\usage{
SmoCurve_decade(vec)
}
\arguments{
\item{vec}{vector of values that the user would like to have smoothed into a curve}
}
\value{
expanded vector of values that determine the shape of the smoothing spline
}
\description{
This function creates a smooth curve from a vector of values.
}
