% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{sutherland_lLik}
\alias{sutherland_lLik}
\title{#'Likelihood of Sutherland Estimates
#'across the tb_progression groups
#'average at the end
#'@param Par_list list (Mpfast,Mrslow,rfast,rRecov)
#'@param n_red
#'@return likelihood}
\usage{
sutherland_lLik(Par_list, N_red = 4)
}
\description{
#'Likelihood of Sutherland Estimates
#'across the tb_progression groups
#'average at the end
#'@param Par_list list (Mpfast,Mrslow,rfast,rRecov)
#'@param n_red
#'@return likelihood
}
