% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_tabby_calib.R
\name{tabby_calib_graphs}
\alias{tabby_calib_graphs}
\title{format the plot
plot the model data
multiply raw output by 1,000 to convert from millions to thousands
reported data for comparison
plot text
# diagnosed cases after 2000 ---------------------------------------------------}
\usage{
tabby_calib_graphs(loc, out_i)
}
\description{
#' graph of total diagnosed cases
#' by total population, US born population, and non-US born population
V0 <- df[51:67,"NOTIF_ALL"]+df[51:67,"NOTIF_MORT_ALL"] #total population
V1 <- df[51:67,"NOTIF_US"]+df[51:67,"NOTIF_MORT_US"]   #US born population
V2 <- df[51:67,"NOTIF_F1"]+df[51:67,"NOTIF_F2"]+df[51:67,"NOTIF_MORT_F1"]+df[51:67,"NOTIF_MORT_F2"]   #non-US born population
}
\details{
#'format the plot
dev.control(displaylist ="enable")

plot.new()
plot(1,1,ylim=c(2,20),xlim=c(2000,2016),xlab="",ylab="",axes=F, log = "y")
axis(1);axis(2,las=2);box()
abline(h=axTicks(2),col="grey85")

#'plot the model data
#'multiply raw output by 1,000 to convert from millions to thousands
lines(2000:2016,V0*1e3,lwd=3,col="white"); lines(2000:2016,V0*1e3,lwd=2,col=1) #total population
lines(2000:2016,V1*1e3,lwd=3,col="white"); lines(2000:2016,V1*1e3,lwd=2,col=4) #US born population
lines(2000:2016,V2*1e3,lwd=3,col="white"); lines(2000:2016,V2*1e3,lwd=2,col=3) #non-US born population

#'reported data for comparison
points(CalibDat[["tot_cases"]][48:64,1],(CalibDat[["tot_cases"]][48:64,2])*1e3,pch=19,cex=0.3) #total population
lines(CalibDat[["tot_cases"]][48:64,1],(CalibDat[["tot_cases"]][48:64,2])*1e3,lty=3,col=1)

notif_fb      <- cbind(CalibDat[["fb_cases"]][,2],1-CalibDat[["fb_cases"]][,2])*CalibDat[["fb_cases"]][,3]
notif_fb <-notif_fb/1000

points(2000:2016,notif_fb[8:24,2],pch=19,cex=0.3,col=4) #US born population
lines(2000:2016,notif_fb[8:24,2],pch=19,lty=3,col=4)

points(2000:2016,notif_fb[8:24,1],pch=19,cex=0.3,col=3) #non-US born population
lines(2000:2016,notif_fb[8:24,1],lty=3,col=3)

#'plot text
mtext("Year",1,2.5,cex=0.9)
mtext("Total TB Cases Identified (000s), 2000-2014",3,.8,font=2,cex=0.8)
legend("topright",c("Reported data (all)","Reported data (US born)","Reported data (foreign born)",
                    "Model (all)","Model (US born)","Model (foreign born)"),
       pch=c(19,19,19,NA,NA,NA),lwd=c(1,1,1,2,2,2),lty=c(3,3,3,1,1,1),col=c(1,4,3,1,4,3),bg="white",ncol=2,cex=.8,pt.cex=0.4)
p <- recordPlot()

saveRDS(p, paste0('utilities/inst/calibration_plots/',loc,'diagnosed_cases_2000.rds'))
format the plot
plot the model data
reported data for comparison
plot text
format the plot
plot the model data
reported data for comparison
plot text
format the plot
plot the model data
reported data for comparison
plot text
format the plot
plot the model data
reported data for comparison
plot text
format the plot
plot the model data
reported data for comparison
plot text
format the plot
plot the model data
reported data for comparison
plot text
format the plot
plot the model data
reported data for comparison
plot text
format the plot
#' graph of total diagnosed cases
#' by total population, US born population, and non-US born population
V0 <- df[51:67,"NOTIF_ALL"]+df[51:67,"NOTIF_MORT_ALL"] #total population
V1 <- df[51:67,"NOTIF_US"]+df[51:67,"NOTIF_MORT_US"]   #US born population
V2 <- df[51:67,"NOTIF_F1"]+df[51:67,"NOTIF_F2"]+df[51:67,"NOTIF_MORT_F1"]+df[51:67,"NOTIF_MORT_F2"]   #non-US born population

#'format the plot
dev.control(displaylist ="enable")

plot.new()
plot(1,1,ylim=c(2,20),xlim=c(2000,2016),xlab="",ylab="",axes=F, log = "y")
axis(1);axis(2,las=2);box()
abline(h=axTicks(2),col="grey85")

#'plot the model data
#'multiply raw output by 1,000 to convert from millions to thousands
lines(2000:2016,V0*1e3,lwd=3,col="white"); lines(2000:2016,V0*1e3,lwd=2,col=1) #total population
lines(2000:2016,V1*1e3,lwd=3,col="white"); lines(2000:2016,V1*1e3,lwd=2,col=4) #US born population
lines(2000:2016,V2*1e3,lwd=3,col="white"); lines(2000:2016,V2*1e3,lwd=2,col=3) #non-US born population

#'reported data for comparison
points(CalibDat[["tot_cases"]][48:64,1],(CalibDat[["tot_cases"]][48:64,2])*1e3,pch=19,cex=0.3) #total population
lines(CalibDat[["tot_cases"]][48:64,1],(CalibDat[["tot_cases"]][48:64,2])*1e3,lty=3,col=1)

notif_fb      <- cbind(CalibDat[["fb_cases"]][,2],1-CalibDat[["fb_cases"]][,2])*CalibDat[["fb_cases"]][,3]
notif_fb <-notif_fb/1000

points(2000:2016,notif_fb[8:24,2],pch=19,cex=0.3,col=4) #US born population
lines(2000:2016,notif_fb[8:24,2],pch=19,lty=3,col=4)

points(2000:2016,notif_fb[8:24,1],pch=19,cex=0.3,col=3) #non-US born population
lines(2000:2016,notif_fb[8:24,1],lty=3,col=3)

#'plot text
mtext("Year",1,2.5,cex=0.9)
mtext("Total TB Cases Identified (000s), 2000-2014",3,.8,font=2,cex=0.8)
legend("topright",c("Reported data (all)","Reported data (US born)","Reported data (foreign born)",
                    "Model (all)","Model (US born)","Model (foreign born)"),
       pch=c(19,19,19,NA,NA,NA),lwd=c(1,1,1,2,2,2),lty=c(3,3,3,1,1,1),col=c(1,4,3,1,4,3),bg="white",ncol=2,cex=.8,pt.cex=0.4)
p <- recordPlot()

saveRDS(p, paste0('utilities/inst/calibration_plots/',loc,'diagnosed_cases_2000.rds'))
format the plot
format the plot
format the plot
format the plot
format the plot
format the plot
}
