% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabby_load.R
\name{tabby_load}
\alias{tabby_load}
\title{loads data for the geography of interest}
\usage{
tabby_load(loc = "US")
}
\arguments{
\item{loc}{two letter postal abbreviation for states; US for national}
}
\value{
env
create a new environment
load necessary datasets into the environment
}
\description{
loads data for the geography of interest
}
