% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_tabby_results.R
\name{tabby_results}
\alias{tabby_results}
\title{gather the outputs from that model run}
\usage{
tabby_results(loc, startyr = 1950, endyr = 2050, Int1 = 0, Int2 = 0,
  Int3 = 0, Int4 = 0, Int5 = 0, Scen1 = 0, Scen2 = 0, Scen3 = 0)
}
\description{
gather the outputs from that model run
}
