% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{tb_deaths_lLik}
\alias{tb_deaths_lLik}
\title{Total TB DEATHS 1999-2014
Motivation: overdispersed poisson, modelled with negbin with overdispersion param = 100 *wts[50:65]}
\usage{
tb_deaths_lLik(V, sgsq = 50)
}
\arguments{
\item{V}{TB deaths by age 1999-2013  (row=15 years, col= 11 ages)}
}
\value{
likelihood
}
\description{
Total TB DEATHS 1999-2014
Motivation: overdispersed poisson, modelled with negbin with overdispersion param = 100 *wts[50:65]
}
