% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{tb_dth_age_lLik}
\alias{tb_dth_age_lLik}
\title{TB DEATHS AGE DISTRIBUTION 1999-2014
Motivation: dirichlet-multinomial, multinomial data with additional non-sampling biases}
\usage{
tb_dth_age_lLik(V, rho = 0.01)
}
\arguments{
\item{V}{table of deaths by age 1999-2014 (row=16 years, col=11 ages)}

\item{rho}{correlation parameter}
}
\value{
likelihood
}
\description{
TB DEATHS AGE DISTRIBUTION 1999-2014
Motivation: dirichlet-multinomial, multinomial data with additional non-sampling biases
}
