% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{tb_graph_all}
\alias{tb_graph_all}
\title{creates plots of various outputs of the model
This script plots a subset of the outputs of the model
and writes these results to a .pdf file.
This script is primarily useful for evaluating the model
output for logic and for identifying potential bugs.
Create a function to be run on a specific model run output to
create simple graphs of all the output for a selected year range}
\usage{
tb_graph_all(start_yr, end_yr, df)
}
\arguments{
\item{start_yr}{year to start the graphs}

\item{end_yr}{year to end the graphs}

\item{df}{dataframe of output for all years}
}
\value{
.pdf of the graphs
}
\description{
creates plots of various outputs of the model
This script plots a subset of the outputs of the model
and writes these results to a .pdf file.
This script is primarily useful for evaluating the model
output for logic and for identifying potential bugs.
Create a function to be run on a specific model run output to
create simple graphs of all the output for a selected year range
}
