% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_demomod.R
\name{tb_graph_demomod}
\alias{tb_graph_demomod}
\title{Create a function to be run on a specific model run output to
create simple graphs of all the output for a selected year range}
\usage{
tb_graph_demomod(df)
}
\arguments{
\item{df}{dataframe of output for all years}

\item{dist}{boolean, if you want generic risk group functions}
}
\value{
.pdf of the graphs
}
\description{
Create a function to be run on a specific model run output to
create simple graphs of all the output for a selected year range
}
