% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{tb_graph_specific}
\alias{tb_graph_specific}
\title{Create a function to be run on a specific model run output to
create easy to run graphs of the output}
\usage{
tb_graph_specific(start_yr, end_yr, df, output)
}
\arguments{
\item{start_yr}{year to start the graphs}

\item{end_yr}{year to end the graphs}

\item{df}{dataframe of output for all years}

\item{output}{character string of output desired}
}
\value{
.pdf of the graphs
}
\description{
Create a function to be run on a specific model run output to
create easy to run graphs of the output
}
