% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{tb_graph_vital}
\alias{tb_graph_vital}
\title{Create a function to be run that will graph the most important outputs
often sums for a designated year range.}
\usage{
tb_graph_vital(start_yr, end_yr, df)
}
\arguments{
\item{start_yr}{year to start the graphs}

\item{end_yr}{year to end the graphs}

\item{df}{dataframe of output for all years}
}
\value{
.pdf of the graphs
}
\description{
Create a function to be run that will graph the most important outputs
often sums for a designated year range.
}
