% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_tb.R
\name{tbdyn_graphs}
\alias{tbdyn_graphs}
\title{creates plots specific to the TB dynamics of the model
and writes these results to a .pdf file.
Use to check outputs of the model; not for distribution
Create a function to be run on a specific mdoel run output to
create simple graphs.}
\usage{
tbdyn_graphs(df)
}
\arguments{
\item{df}{dataframe of output for all years}
}
\value{
.pdf of all graphs
}
\description{
creates plots specific to the TB dynamics of the model
and writes these results to a .pdf file.
Use to check outputs of the model; not for distribution
Create a function to be run on a specific mdoel run output to
create simple graphs.
}
