% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{tiemersma_lLik}
\alias{tiemersma_lLik}
\title{Likelihood of Tiemersa Estimates}
\usage{
tiemersma_lLik(Par)
}
\arguments{
\item{Par}{vector (pfast,rslow,rfast,rRecov)}
}
\value{
likelihood
}
\description{
Likelihood of Tiemersa Estimates
}
