% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{tltbi_dist_lLik}
\alias{tltbi_dist_lLik}
\title{DISTRIBUTION OF LTBI TREATMENT INITS 2002}
\usage{
tltbi_dist_lLik(V)
}
\arguments{
\item{V}{dist TLTBI inits in 2002 (vector fraction FB, HR in 2002)}
}
\value{
likelihood
}
\description{
DISTRIBUTION OF LTBI TREATMENT INITS 2002
}
