% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{tltbi_tot_lLik}
\alias{tltbi_tot_lLik}
\title{TOTAL LTBI TREATMENT INITS 2002
Motivation: norm, mean centered with CI = +/- 10% of mean}
\usage{
tltbi_tot_lLik(V)
}
\arguments{
\item{V}{scalar total TLTBI inits in 2002}
}
\value{
likelihood
}
\description{
TOTAL LTBI TREATMENT INITS 2002
Motivation: norm, mean centered with CI = +/- 10% of mean
}
