% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{tot_pop16_ag_fb_lLik}
\alias{tot_pop16_ag_fb_lLik}
\title{TOTAL POP AGE DISTRIBUTION 2014
Motivation: reported estimates represent pseudo-data for a multinomial likelihood, with ESS = 500}
\usage{
tot_pop16_ag_fb_lLik(V, ESS = 500)
}
\arguments{
\item{V}{US pop in 2016 (row=11 ages, col= rec fb, fb)}

\item{ESS}{explained sum of squares}
}
\value{
likelihood
}
\description{
TOTAL POP AGE DISTRIBUTION 2014
Motivation: reported estimates represent pseudo-data for a multinomial likelihood, with ESS = 500
}
