% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{tot_pop_age_lLik}
\alias{tot_pop_age_lLik}
\title{#' TOTAL POP AGE DISTRIBUTION  NO NATIVITY (ONLY FOR USE IN DEMO MODEL) 1999-2014
#'@param V table of POP by age 2016 (row=16 years, col=11 ages)
#'@param rho correlation parameter
#'@return likelihood}
\usage{
tot_pop_age_lLik(V, ESS = 500)
}
\description{
#' TOTAL POP AGE DISTRIBUTION  NO NATIVITY (ONLY FOR USE IN DEMO MODEL) 1999-2014
#'@param V table of POP by age 2016 (row=16 years, col=11 ages)
#'@param rho correlation parameter
#'@return likelihood
}
