% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{tot_pop_yr_fb_lLik}
\alias{tot_pop_yr_fb_lLik}
\title{TOTAL POP EACH DECADE, FOR FB
Motivation: norm, mean centered with CI = +/- 2 million wts[1+0:6*10]}
\usage{
tot_pop_yr_fb_lLik(V)
}
\arguments{
\item{V}{total pop (rows=year, cols=us, fb)}
}
\value{
likelihood
}
\description{
TOTAL POP EACH DECADE, FOR FB
Motivation: norm, mean centered with CI = +/- 2 million wts[1+0:6*10]
}
