% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{tx_outcomes_lLik}
\alias{tx_outcomes_lLik}
\title{TREATMENT OUTCOMES 1993-2014
Motivation: dirichlet-multinomial, multinomial data with additional non-sampling biases}
\usage{
tx_outcomes_lLik(V, rho = 0.05)
}
\arguments{
\item{V}{table of treatment outcomes 1993-2012 (row=20 years, col= complete, discontinue, dead)}

\item{rho}{correlation parameter}
}
\value{
likelihood
}
\description{
TREATMENT OUTCOMES 1993-2014
Motivation: dirichlet-multinomial, multinomial data with additional non-sampling biases
}
