% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_univ.r
\name{UnivOptim}
\alias{UnivOptim}
\title{This function is for univariate optimization of the parameters
used in tb_model.cpp}
\usage{
UnivOptim(parr)
}
\arguments{
\item{parr}{parameter vector}
}
\value{
list
}
\description{
This function is for univariate optimization of the parameters
used in tb_model.cpp
}
