% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{US_dth_tot_lLik}
\alias{US_dth_tot_lLik}
\title{TOTAL US DEATHS
1970,1975,1980,1985,1990-2007
Motivation: norm, mean centered with CI = +/- 5% of mean}
\usage{
US_dth_tot_lLik(V)
}
\value{
likelihood
}
\description{
TOTAL US DEATHS
1970,1975,1980,1985,1990-2007
Motivation: norm, mean centered with CI = +/- 5% of mean
}
