% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions_st.r
\name{US_notif_tot_lLik_st}
\alias{US_notif_tot_lLik_st}
\title{Total Diagnosed US Cases 1993-2017
Motivation: Normal, mean centered with CI = +/- 5% of the mean}
\usage{
US_notif_tot_lLik_st(V, st)
}
\arguments{
\item{V}{vector of total notifications 1953-2014}
}
\value{
likelihood
}
\description{
Total Diagnosed US Cases 1993-2017
Motivation: Normal, mean centered with CI = +/- 5% of the mean
}
