% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{US_pop_tot_lLik}
\alias{US_pop_tot_lLik}
\title{TOTAL US POP No nativity (ONLY FOR USE IN DEMO MODEL)
1970,1975,1980,1985,1990-2007
Motivation: norm, mean centered with CI = +/- 5% of mean}
\usage{
US_pop_tot_lLik(X)
}
\arguments{
\item{X}{vector of total deaths in US from 1971-2016, fraction of millions}
}
\value{
j likelihood
}
\description{
TOTAL US POP No nativity (ONLY FOR USE IN DEMO MODEL)
1970,1975,1980,1985,1990-2007
Motivation: norm, mean centered with CI = +/- 5% of mean
}
