% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_demo.R
\docType{data}
\name{US_tot_mort}
\alias{US_tot_mort}
\title{creates plots of the demographics of the model
and writes these results to a .pdf file.
Use to check both with and without tb in the model}
\format{An object of class \code{data.frame} with 22 rows and 2 columns.}
\usage{
US_tot_mort
}
\description{
creates plots of the demographics of the model
and writes these results to a .pdf file.
Use to check both with and without tb in the model
}
\keyword{datasets}
