% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_weightedmortality.R
\name{weight_mort}
\alias{weight_mort}
\title{THE CODE BELOW SOURCES THE MODEL INPUTS AND THEN FORMATS THEM FOR USE
IN THE OUTPUTSINTZ FUNCTION THAT CALLS CSIM FROM THE TB_MODEL.CPP
FUNCTION FILE. ALL VARIABLE NAMES THAT END IN t ARE INDEXED BY TIME
VARIABLE NAMES BEGINNING WITH m ARE MATRICES & V ARE VECTORS.}
\usage{
weight_mort(loc)
}
\arguments{
\item{loc}{two digit mailing abbreviation of state}
}
\value{
matrix of denominators 10x1201
}
\description{
THE CODE BELOW SOURCES THE MODEL INPUTS AND THEN FORMATS THEM FOR USE
IN THE OUTPUTSINTZ FUNCTION THAT CALLS CSIM FROM THE TB_MODEL.CPP
FUNCTION FILE. ALL VARIABLE NAMES THAT END IN t ARE INDEXED BY TIME
VARIABLE NAMES BEGINNING WITH m ARE MATRICES & V ARE VECTORS.
}
