% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_reblnc_pop.R
\docType{data}
\name{wgt_targ}
\alias{wgt_targ}
\title{Define weights of population in the target group and gen pop}
\format{An object of class \code{numeric} of length 1.}
\usage{
wgt_targ
}
\description{
Define weights of population in the target group and gen pop
}
\keyword{datasets}
