% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R, R/model_load.R
\docType{data}
\name{wts}
\alias{wts}
\title{Log-likelihood functions
Assign the calibration importance weights from CalibDat
These weights are based on year of the simulation.}
\format{An object of class \code{numeric} of length 67.}
\usage{
wts

wts
}
\description{
Log-likelihood functions
Assign the calibration importance weights from CalibDat
These weights are based on year of the simulation.

format the calibration data
}
\keyword{datasets}
