% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{bspline}
\alias{bspline}
\title{This function is used to return the basis of a spline function.}
\usage{
bspline(x, k, i, m)
}
\arguments{
\item{x}{param 1}

\item{k}{knots}

\item{i}{param 2}

\item{m}{param 3}
}
\value{
basis of a spline
}
\description{
This function is used to return the basis of a spline function.
}
