% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib.R
\name{calib}
\alias{calib}
\title{if any output is missing or negative or if any model state population is negative
set the likelihood to a hugely negative number (penalized)}
\usage{
calib(samp_i, ParMatrix)
}
\description{
if any output is missing or negative or if any model state population is negative
set the likelihood to a hugely negative number (penalized)
}
