% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_calib.R
\name{calib_graphs}
\alias{calib_graphs}
\title{load the parameter data set & model inputs
graph of total diagnosed cases
by total population, US born population, and non-US born population
format the plot
plot the model data
multiply raw output by 1,000 to convert from millions to thousands
reported data for comparison
plot text
graph of total diagnosed cases
by total population, US born population, and non-US born population
format the plot
plot the model data
multiply raw output by 1,000 to convert from millions to thousands
reported data for comparison
plot text
Percent of Total Cases Non-US Born Population
format the plot
plot the model data
reported data for comparison
plot text
Percent of Non-US Born Cases from Recent Immigrant Population
format the plot
plot the model data
reported data for comparison
plot text
Age distribution of Cases
0-24 yrs, 25-44 yrs, 45-64 yrs, 65+ yrs
format the plot
plot the model data
reported data for comparison
plot text
Age Distribution of TB Cases in Percentages
0-24 yrs, 25-44 yrs, 45-64 yrs, 65+ yrs}
\usage{
calib_graphs(df)
}
\description{
V   <- (df[51:65,136:146]+df[51:65,189:199])
  V2  <- V[,-11]
  V2[,10] <- V2[,10]+V[,11]
}
\details{
#'format the plot
  plot(0,0,ylim=c(0,max(range(V2))),xlim=c(0.6,10.4),xlab="",ylab="",axes=F,col=NA)
  axis(1,1:10,paste(c("0-4",paste(0:7*10+5,1:8*10+4,sep="-"),"85+"),"\nyears",sep=""),
       tick=F,cex.axis=0.6)
  axis(1,1:11-0.5,rep("",11))
  axis(2,las=2);box()
  abline(h=axTicks(2),col="grey85")

#'plot the model data
  for(i in 1:10) polygon(i+c(-.5,.5,.5,-.5),c(0,0,V2[i],V2[i]),border="white",col="lightblue")

#'reported data for comparison
  points(1:10,colSums(notif_age[7:21,])/sum(notif_age[7:21,])*100,pch=19,cex=1.2)

#'plot text
  mtext("Age Group",1,2.5,cex=0.9)
  mtext("Age Distribution of TB Cases (%), 2000-14",3,.8,font=2,cex=0.8)
  legend("topright",c("Reported data","Model"),pch=c(19,15),lwd=NA,
         pt.cex=c(1,2),col=c("black","lightblue"),bg="white")
Distribution of Cases across the TB Progression Categories
format the plot
plot the model data
reported data for comparison
plot text
Treatment Outcomes 1993-2014
format the plot
plot the model data
reported data for comparison
plot text
LTBI Prevalance by Age in 2011, US born
format the plot
plot the model data
reported data for comparison
plot text
LTBI Prevalance by Age in 2011, non-US born
format the plot
plot the model data
reported data for comparison
plot text
Age Distribution of TB Deaths 1999-2013
format the plot
plot the model data
reported data for comparison
plot text
}
