% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cSim}
\alias{cSim}
\title{This is the simulation model code.}
\usage{
cSim(
  nYrs,
  nRes,
  rDxt,
  TxQualt,
  InitPop,
  Mpfast,
  ExogInf,
  MpfastPI,
  Mrslow,
  rrSlowFB,
  rfast,
  RRcurDef,
  rSlfCur,
  p_HR,
  vTMort,
  RRmuRF,
  RRmuHR,
  Birthst,
  HrEntEx,
  ImmNon,
  ImmLat,
  ImmAct,
  ImmFst,
  net_mig_usb,
  net_mig_nusb,
  mubt,
  RelInf,
  RelInfRg,
  Vmix,
  rEmmigFB,
  TxVec,
  TunTxMort,
  rDeft,
  rLtScrt,
  rrTestHr,
  rrTestLrNoTb,
  LtTxPar,
  ttt_samp_dist,
  ttt_ag,
  ttt_na,
  ttt_month,
  ttt_pop_frc,
  ttt_ltbi,
  LtDxPar_lt,
  LtDxPar_nolt,
  RRdxAge,
  rRecov,
  pImmScen,
  EarlyTrend,
  pReTx,
  ag_den,
  NixTrans,
  dist_gen,
  trans_mat_tot_ages
)
}
\arguments{
\item{nYrs}{number of years to run the model.}

\item{nRes}{number of results of the model}

\item{rDxt}{Rate of active TB diagnosis over time}

\item{TxQualt}{active TB treatment over time}

\item{InitPop}{Initial Population matrix}

\item{Mpfast}{Matrix of the probabilities of fast TB progression (age x TB prog risk group)}

\item{ExogInf}{Exogenous infection risk for non-US born population}

\item{MpfastPI}{Matrix of the probabilities of fast TB progression w/ partial immunity (age x TB prog risk group)}

\item{Mrslow}{matrix of the rates of slow progression (age x TB prog risk group)}

\item{rrSlowFB}{rate of fast TB progression}

\item{RRcurDef}{Rate Ratio for cure given treatment defaul}

\item{rSlfCur}{rate of self cure from active TB}

\item{p_HR}{probability of high risk population @ entry into the model}

\item{vTMort}{vector of TB mortality rates}

\item{RRmuRF}{rate ratio of mortality across mortality risk group}

\item{RRmuHR}{rate ratio of mortality across low/high risk dimension}

\item{Birthst}{Births over time}

\item{HrEntEx}{Matrix of Entry and Exit rates into the High Risk population}

\item{ImmNon}{Immigration with no TB}

\item{ImmLat}{Immigration with Latent TB}

\item{ImmAct}{Immigration with Active TB}

\item{ImmFst}{Immigration with Fast Progressing TB}

\item{net_mig_usb}{net internal migration usb}

\item{net_mig_nusb}{net internal migration nusb}

\item{mubt}{background mortality over time}

\item{RelInf}{beta}

\item{RelInfRg}{beta based off of risk group}

\item{Vmix}{1-sigma}

\item{rEmmigFB}{rate of emmigration in non-US born population}

\item{TxVec}{vector of parameters for TB Tx}

\item{TunTxMort}{Tuning parameter for mortality on TB Tx}

\item{rDeft}{rate of default from TB treatment over time}

\item{rLtScrt}{rate of latent screening over time}

\item{rrTestHr}{RR of latent testing for HR}

\item{LtTxPar}{matrix of latent treatment parameters}

\item{ttt_samp_dist}{probabilities of screening for ttt intervention}

\item{ttt_ag}{which age groups to apply ttt}

\item{ttt_na}{which nativity groups to apply ttt}

\item{ttt_month}{when to apply ttt interventions}

\item{ttt_pop_frc}{population size to apply the ltbi prev to}

\item{ttt_ltbi}{how much to increase ltbi}

\item{LtDxPar_lt}{matrix of latent diagnosis parameters}

\item{LtDxPar_nolt}{matrix of latent diagnosis parameters}

\item{RRdxAge}{vector of rate ratios for TB diagnosis by age}

\item{rRecov}{rate of recovery from latent slow to safe tb state}

\item{pImmScen}{lack of reactivitiy to IGRA for Sp}

\item{EarlyTrend}{ramp down of TB in burn-in}

\item{pReTx}{probability of re-treatment for TB}

\item{ag_den}{denominator used in the aging process}

\item{NixTrans}{reduction of transmission over time}

\item{dist_gen}{general distribution across tb progression and mort}

\item{rrTestNoTb}{RR of latent testing for those without TB}
}
\value{
Outputs a list of outputs
}
\description{
runs a simulation of the tb model
}
