% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{dDirMult}
\alias{dDirMult}
\title{Dirichlet multinomial density function}
\usage{
dDirMult(M, n, Rho)
}
\arguments{
\item{M}{params of the dirichlet, i.e. model strain distribution}

\item{n}{category counts from the survey}

\item{rho}{correlation parameter = 1/sample size}
}
\value{
M density
}
\description{
Dirichlet multinomial density function
}
