% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention_programchangedefault.R
\name{def_prgchng}
\alias{def_prgchng}
\title{Function that will return default values of the program change values.}
\usage{
def_prgchng(ParVec)
}
\arguments{
\item{ParVec}{parameter vector}
}
\value{
vector of the default values
}
\description{
Function that will return default values of the program change values.
}
