% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_reblnc_pop.R
\docType{data}
\name{diff_i_v}
\alias{diff_i_v}
\title{Create several vectors populated with NA to input into the model;}
\format{An object of class \code{logical} of length 16.}
\usage{
diff_i_v
}
\description{
Create several vectors populated with NA to input into the model;
}
\keyword{datasets}
