% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{ferebee_lLik}
\alias{ferebee_lLik}
\title{Likelihood of Ferebee Estimates
across the tb_progression groups
average at the end}
\usage{
ferebee_lLik(Par_list, N_red = 4)
}
\arguments{
\item{Par_list}{list (pfast,rslow,rfast,rRecov)}

\item{n_red}{crude reductions the likelihood (allows non-sampling bias)}
}
\value{
likelihood
}
\description{
Likelihood of Ferebee Estimates
across the tb_progression groups
average at the end
}
