% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{getmode}
\alias{getmode}
\title{Simple function to return the mode of a vector}
\usage{
getmode(v)
}
\arguments{
\item{x}{some vector (character or numeric)}
}
\value{
mode
}
\description{
Simple function to return the mode of a vector
}
