% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMIS_noRB.R
\name{llikelihood_noRB}
\alias{llikelihood_noRB}
\title{Local parallelization via multicore}
\usage{
llikelihood_noRB(opt_mat, loc, n_cores = 1)
}
\arguments{
\item{opt_mat}{matrix of parameters}

\item{loc}{two digit code of location}

\item{n_cores}{number of cores to use on the cluster}
}
\value{
lLik
}
\description{
Local parallelization via multicore
}
