% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMIS_functions.R
\name{llikelihoodZ}
\alias{llikelihoodZ}
\title{This scrprmst creates a function that loops over the log-likelihood
functions found in calib_functions.R and updates the
this is the function that goes into the optimizer}
\usage{
llikelihoodZ(samp_i, start_mat)
}
\arguments{
\item{samp_i}{sample id}

\item{start_mat}{matrix of parameters  # Par = par_1}
}
\value{
lLik
if any output is missing or negative or if any model state population is negative
set the likelihood to a hugely negative number (penalized)
TOTAL DIAGNOSED CASES 1953-2016 - index is same
US CASES AGE DISTRIBUTION 1993-2016 - index updated
FB CASES AGE DISTRIBUTION 1993-2016 - index updated
CASES FB DISTRIBUTION 1993-2016 - index updated
CASES FB, US 2012-2016  SLOPE - index updated
CASES HR DISTRIBUTION 1993-2016 - index updated
high risk first column, low risk second column
CASES FB RECENT ENTRY DISTRIBUTION 1993-2014 index updated
recent immigrants column one; long term in column two
TREATMENT OUTCOMES 1993-2014 - index updated
TOTAL LTBI TREATMENT INITS 2002 - index updated
DIST LTBI TREATMENT INITS 2002 - index updated
LTBI PREVALENCE BY AGE 2011, US - index updated
LTBI PREVALENCE BY AGE 2011, FB - index updated
TOTAL POP EACH DECADE, BY US/FB - index updated (maybe)
TOTAL POP AGE DISTRIBUTION 2016 index updated
TOTAL DEATHS WITH TB 1999-2014 - index updated
TB DEATHS 1999-2014 BY AGE - index updated above
#' Total DEATHS 2017
#' Total DEATHS 1999-2016 BY AGE
#' Mort_dist 2016
#' Mort_dist 2016
HOMELESS POP 2010 - index updated
LIKELIHOOD FOR BORGDORFF, FEREBEE & SUTHERLAND ESTIMATES
}
\description{
This scrprmst creates a function that loops over the log-likelihood
functions found in calib_functions.R and updates the
this is the function that goes into the optimizer
}
