% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_IMIS_noTB.R
\name{llikelihoodZ_noTB}
\alias{llikelihoodZ_noTB}
\title{This scrprmst creates a function that loops over the log-likelihood
functions found in calib_functions.R and updates the
this is the function that goes into the optimizer}
\usage{
llikelihoodZ_noTB(samp_i, opt_mat)
}
\arguments{
\item{samp_i}{sample id}

\item{opt_mat}{matrix of parameters  # Par = par_1}
}
\value{
lLik
format P
if any output is missing or negative or if any model state population is negative
set the likelihood to a hugely negative number (penalized)
TOTAL POP EACH DECADE, BY US/FB - index updated (maybe)
v17  <- M[,31]+M[,32]
addlik <- tot_pop_yr_fb_lLik(V=v17); addlik
lLik <- lLik + addlik
#' TOTAL POP AGE DISTRIBUTION 2016 index updated
v18  <- cbind(M[67,33:43],M[67,44:54])
addlik <- tot_pop16_ag_fb_lLik(V=v18); addlik
lLik <- lLik + addlik
Total DEATHS 1999-2016 BY AGE
#' Mort_dist 2016
}
\description{
This scrprmst creates a function that loops over the log-likelihood
functions found in calib_functions.R and updates the
this is the function that goes into the optimizer
}
