% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMIS_functions_st.r
\name{llikelihoodZ_st}
\alias{llikelihoodZ_st}
\title{if any output is missing or negative or if any model state population is negative
LTBI PREVALENCE BY AGE 2011, FB - index updated
LIKELIHOOD FOR BORGDORFF, FEREBEE & SUTHERLAND ESTIMATES
Total DEATHS 1979-2016
#' #' Total DEATHS 2015-2016 BY AGE}
\usage{
llikelihoodZ_st(samp_i, ParMatrix, loc = "MA")
}
\description{
#' #' Mort_dist 2016 dirchlet
}
