% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMIS_functions_st.r
\name{llikelihoodZ_st}
\alias{llikelihoodZ_st}
\title{Total DEATHS 1979-2016
#' Total DEATHS 1999-2016 BY AGE
v20b  <- M[50:67,121:131]
addlik <- tot_dth_age_lLik(V=v20b); addlik
lLik <- lLik + addlik
#' Mort_dist 2016
LIKELIHOOD FOR BORGDORFF, FEREBEE & SUTHERLAND ESTIMATES}
\usage{
llikelihoodZ_st(samp_i, ParMatrix, st)
}
\description{
Total DEATHS 1979-2016
#' Total DEATHS 1999-2016 BY AGE
v20b  <- M[50:67,121:131]
addlik <- tot_dth_age_lLik(V=v20b); addlik
lLik <- lLik + addlik
#' Mort_dist 2016
LIKELIHOOD FOR BORGDORFF, FEREBEE & SUTHERLAND ESTIMATES
}
