% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMIS_functions_st.r
\name{llikelihoodZ_st}
\alias{llikelihoodZ_st}
\title{if any output is missing or negative or if any model state population is negative
LTBI PREVALENCE BY AGE 2011, NUSB - index updated
LIKELIHOOD FOR BORGDORFF, FEREBEE & SUTHERLAND ESTIMATES
### ### ### HOMELESS POP 2010  ### ### ### ### ### ###
#' Total DEATHS 2016
#' #' #' Total DEATHS 2015-2016 BY AGE
#' #' Mort_dist 2016 dirchlet}
\usage{
llikelihoodZ_st(samp_i, ParMatrix, loc, TB = 1)
}
\description{
if any output is missing or negative or if any model state population is negative
LTBI PREVALENCE BY AGE 2011, NUSB - index updated
LIKELIHOOD FOR BORGDORFF, FEREBEE & SUTHERLAND ESTIMATES
### ### ### HOMELESS POP 2010  ### ### ### ### ### ###
#' Total DEATHS 2016
#' #' #' Total DEATHS 2015-2016 BY AGE
#' #' Mort_dist 2016 dirchlet
}
