% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabby2_allscenarios.R
\name{make_all_scenarios}
\alias{make_all_scenarios}
\title{THIS FUNCTION INPUTS A TABLE OF PARAMETERS AND RUNS THE TB MODEL
AND GENERATES AN ARRAY OF OUTPUTS FOR EACH POSSIBLE PREDEFINED
SCENARIO; THIS ARRAY IS THEN SAVED TO THE DATA FOLDER FOR THAT
LOCATION}
\usage{
make_all_scenarios(loc, ParMatrix)
}
\arguments{
\item{loc}{two character location code for the model location}

\item{ParMatrix}{parameters to use in the simulation}
}
\description{
THIS FUNCTION INPUTS A TABLE OF PARAMETERS AND RUNS THE TB MODEL
AND GENERATES AN ARRAY OF OUTPUTS FOR EACH POSSIBLE PREDEFINED
SCENARIO; THIS ARRAY IS THEN SAVED TO THE DATA FOLDER FOR THAT
LOCATION
}
