% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabby2_caliboutputs.R
\name{model_calib_outputs}
\alias{model_calib_outputs}
\title{THIS FUNCTION READS IN THE BASE CASE SCENARIO AND THEN
CREATES INDIVIDUAL RDS FILES FOR OUTPUTS NECESSARY FOR
THE TABBY2 COMPARISON TO RECENT DATA PLOTS}
\usage{
model_calib_outputs(loc = "US", res, simp.date)
}
\arguments{
\item{loc}{two character location code for the model location}

\item{simp.date}{date to append to the files; should match the optim date}

\item{bc.array}{base case results array}

\item{samp.i}{which simulation to use (1:10)}
}
\description{
THIS FUNCTION READS IN THE BASE CASE SCENARIO AND THEN
CREATES INDIVIDUAL RDS FILES FOR OUTPUTS NECESSARY FOR
THE TABBY2 COMPARISON TO RECENT DATA PLOTS
}
