% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_load.R
\name{model_load}
\alias{model_load}
\title{loads data for the geography of interest}
\usage{
model_load(loc = "US")
}
\arguments{
\item{loc}{two letter postal abbreviation for states; US for national}
}
\value{
void
load necessary datasets
Model Input
}
\description{
loads data for the geography of interest
}
