% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_load.R
\name{model_load}
\alias{model_load}
\title{Calibration Target Environment
all other loads will be dependent on geography
lazy load necessary datasets
Model Input
creation of background parameters
elements of P will be replaced from either the StartVals in the case
of optimization or the user inputted dataset}
\usage{
model_load(loc = "US")
}
\description{
format the calibration data
}
