% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{mort_dist_lLik}
\alias{mort_dist_lLik}
\title{Mortality Risk Group Distribution 1999-2014
Motivation: dirichlet-multinomial, multinomial data with additional non-sampling biases}
\usage{
mort_dist_lLik(V, rho = 0.01)
}
\arguments{
\item{V}{table of mort_dist 1999-2014 (row=16 years, col=11 ages)}

\item{rho}{correlation parameter}
}
\value{
likelihood
}
\description{
Mortality Risk Group Distribution 1999-2014
Motivation: dirichlet-multinomial, multinomial data with additional non-sampling biases
}
