% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions_st.r
\name{notif_tot_lLik_st}
\alias{notif_tot_lLik_st}
\title{Calibration Functions for use with the tb_model.cpp model for state level
This script creates several individual log likelihood functions
for the calibration of the State Level TB model in tb_model.cpp
These llikelihood functions are called in IMIS_functions.R
takes in the outputs and calibration data and creates likelihood functions
Total Diagnosed Cases 1953-2016
Motivation: Normal, mean centered with CI = +/- 5% of the mean}
\usage{
notif_tot_lLik_st(V, st)
}
\arguments{
\item{V}{vector of total notifications 1953-2014}
}
\value{
likelihood
}
\description{
Calibration Functions for use with the tb_model.cpp model for state level
This script creates several individual log likelihood functions
for the calibration of the State Level TB model in tb_model.cpp
These llikelihood functions are called in IMIS_functions.R
takes in the outputs and calibration data and creates likelihood functions
Total Diagnosed Cases 1953-2016
Motivation: Normal, mean centered with CI = +/- 5% of the mean
}
