% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_reblnc_pop.R
\docType{data}
\name{num_mRF}
\alias{num_mRF}
\title{Define the number of mortality and progression risk groups}
\format{An object of class \code{numeric} of length 1.}
\usage{
num_mRF
}
\description{
Define the number of mortality and progression risk groups
}
\keyword{datasets}
