% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_optimfunction.R
\name{optim_b}
\alias{optim_b}
\title{This function is used to run a optimization on the input parameters.
The only input to the function is a "batch number" which will determine
which row of the starting values dataframe the optimization will use.
It will return 8 datasets of the optimized parameters --one from each
optimization step. The final (8th) optimization step is an univariate
optimization.
The function requires the use of the ParamInit Rdata file and the
StartValues Rdata file.}
\usage{
optim_b(df, samp_i = 1)
}
\arguments{
\item{df}{dataframe or matrix of starting values data frame}

\item{samp_i}{which rows of the data frame to use}
}
\value{
8 datasets from optimization loop
}
\description{
This function is used to run a optimization on the input parameters.
The only input to the function is a "batch number" which will determine
which row of the starting values dataframe the optimization will use.
It will return 8 datasets of the optimized parameters --one from each
optimization step. The final (8th) optimization step is an univariate
optimization.
The function requires the use of the ParamInit Rdata file and the
StartValues Rdata file.
}
