% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_optimfunction.R
\name{optim_b_st}
\alias{optim_b_st}
\title{This function is used to run a optimization on the input parameters.
The only input to the function is a "batch number" which will determine
which row of the starting values dataframe the optimization will use.
It will return 8 datasets of the optimized parameters --one from each
optimization step. The final (8th) optimization step is an univariate
optimization.
The function requires the use of the ParamInit Rdata file and the
StartValues Rdata file.}
\usage{
optim_b_st(df, samp_i = 1, loc = "MA", TB = 1)
}
\arguments{
\item{df}{dataframe or matrix of starting values data frame}

\item{samp_i}{which rows of the data frame to use}

\item{loc}{USPS code of state}

\item{TB}{boolean for TB likelihoods}
}
\value{
8 datasets from optimization loop
}
\description{
This function is used to run a optimization on the input parameters.
The only input to the function is a "batch number" which will determine
which row of the starting values dataframe the optimization will use.
It will return 8 datasets of the optimized parameters --one from each
optimization step. The final (8th) optimization step is an univariate
optimization.
The function requires the use of the ParamInit Rdata file and the
StartValues Rdata file.
}
