% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_optim_data.R
\name{optim_data}
\alias{optim_data}
\title{The function requires the use of the ParamInit Rdata file and the
StartValues Rdata file.}
\usage{
optim_data(batches, loc, date)
}
\arguments{
\item{batches}{dataframe or matrix of starting values data frame}

\item{loc}{two digit location code}

\item{date}{MM-DD date on optim files}
}
\value{
datafile of all the optimized data parameters
}
\description{
The function requires the use of the ParamInit Rdata file and the
StartValues Rdata file.
}
