% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{param}
\alias{param}
\title{THE CODE BELOW SOURCES THE MODEL INPUTS AND THEN FORMATS THEM FOR USE
IN THE OUTPUTSINTZ FUNCTION THAT CALLS CSIM FROM THE TB_MODEL.CPP
FUNCTION FILE. ALL VARIABLE NAMES THAT END IN t ARE INDEXED BY TIME
VARIABLE NAMES BEGINNING WITH m ARE MATRICES & V ARE VECTORS.}
\usage{
param(PV, loc, prg_chng)
}
\arguments{
\item{loc}{two-digit abbreviation for location}

\item{prg_chng}{vector of program change values}

\item{P}{vector of unformatted parameter values}
}
\value{
Params list
}
\description{
THE CODE BELOW SOURCES THE MODEL INPUTS AND THEN FORMATS THEM FOR USE
IN THE OUTPUTSINTZ FUNCTION THAT CALLS CSIM FROM THE TB_MODEL.CPP
FUNCTION FILE. ALL VARIABLE NAMES THAT END IN t ARE INDEXED BY TIME
VARIABLE NAMES BEGINNING WITH m ARE MATRICES & V ARE VECTORS.
}
