% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_initial.R
\name{param_init}
\alias{param_init}
\title{This function takes the same inputs as the Outputs}
\usage{
param_init(
  PV,
  loc,
  Int1 = 0,
  Int2 = 0,
  Int3 = 0,
  Int4 = 0,
  Int5 = 0,
  Scen1 = 0,
  Scen2 = 0,
  Scen3 = 0,
  Scen4 = 0,
  Scen5 = 0,
  Scen6 = 0,
  prg_chng,
  ttt_list
)
}
\arguments{
\item{PV}{vector of Inputs to format}

\item{loc}{two letter abbreviation}

\item{Int1}{boolean for intervention 1}

\item{Int2}{boolean for intervention 2}

\item{Int3}{boolean for intervention 3}

\item{Int4}{boolean for intervention 4}

\item{Int5}{boolean for intervention 5}

\item{Scen1}{boolean for scenario 1}

\item{Scen2}{boolean for scenario 2}

\item{Scen3}{boolean for scenario 3}

\item{Scen4}{boolean for scenario 4}

\item{Scen5}{boolean for scenario 5}

\item{Scen6}{boolean for scenario 6}

\item{prg_chng}{vector of program change values}

\item{ttt_list}{list of ttt changes}
}
\value{
InputParams list
}
\description{
This function takes the same inputs as the Outputs
}
