% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabby2_makeresults.R
\name{reshape_results}
\alias{reshape_results}
\title{THIS FUNCTION RESHAPES THE 9 INDIVIDUAL FILES INTO THE
FORMAT NEEDED FOR TABBY2 VISUALIZATION}
\usage{
reshape_results(loc, results)
}
\arguments{
\item{loc}{two character location code for the model location}

\item{results}{is a list of results to be reshaped}
}
\description{
THIS FUNCTION RESHAPES THE 9 INDIVIDUAL FILES INTO THE
FORMAT NEEDED FOR TABBY2 VISUALIZATION
}
