% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabby2_makeresults.R
\name{reshape_results}
\alias{reshape_results}
\title{THIS FUNCTION INPUTS A TABLE OF PARAMETERS AND RUNS THE TB MODEL
AND GENERATES AN ARRAY OF OUTPUTS FOR EACH POSSIBLE PREDEFINED
SCENARIO; THIS ARRAY IS THEN SAVED TO THE DATA FOLDER FOR THAT
LOCATION}
\usage{
reshape_results(loc, results.list)
}
\arguments{
\item{loc}{two character location code for the model location}
}
\description{
THIS FUNCTION INPUTS A TABLE OF PARAMETERS AND RUNS THE TB MODEL
AND GENERATES AN ARRAY OF OUTPUTS FOR EACH POSSIBLE PREDEFINED
SCENARIO; THIS ARRAY IS THEN SAVED TO THE DATA FOLDER FOR THAT
LOCATION
}
