% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_sampleprior.R
\name{sample.prior1}
\alias{sample.prior1}
\title{This script defines the functions to sample the prior.
This function samples the prior using a random normal generation.}
\usage{
sample.prior1(n)
}
\arguments{
\item{n}{number of sample sets to generate}
}
\value{
n sample sets from the prior in a single object
}
\description{
This script defines the functions to sample the prior.
This function samples the prior using a random normal generation.
}
