% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_functions.R
\name{tb_dth_tot_lLik}
\alias{tb_dth_tot_lLik}
\title{TOTAL TB DEATHS 1999-2014
Motivation: norm, mean centered with CI = +/- 5% of mean}
\usage{
tb_dth_tot_lLik(V)
}
\value{
likelihood
}
\description{
TOTAL TB DEATHS 1999-2014
Motivation: norm, mean centered with CI = +/- 5% of mean
}
